% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_phenocycler_diffuse_redesign.R
\name{createSCOPE_phenocycler_redesign}
\alias{createSCOPE_phenocycler_redesign}
\title{Create a scope object from an organised PhenoCycler bundle using redesign
entrypoints.}
\usage{
createSCOPE_phenocycler_redesign(
  bundle_dir,
  marker_policy = makePhenoCyclerMarkerPolicy(),
  redesign_config = makePhenoCyclerDiffuseRedesignConfig(),
  adapter_config = makePhenoCyclerRedesignAdapterConfig(),
  promote = adapter_config$promote,
  verbose = TRUE
)
}
\arguments{
\item{bundle_dir}{Path to the organised PhenoCycler bundle.}

\item{marker_policy}{Marker-family policy object.}

\item{redesign_config}{Diffuse redesign config.}

\item{adapter_config}{Redesign adapter config.}

\item{promote}{Which redesigned representation, if any, should be promoted
into `@grid`.}

\item{verbose}{Emit progress messages when TRUE.}
}
\value{
A `scope_object`.
}
\description{
This constructor is explicit opt-in and does not modify the v1
`createSCOPE_phenocycler()` path.
}
