% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_threading.r, R/zzz.r
\name{detectOS}
\alias{detectOS}
\title{Detect operating system}
\usage{
detectOS()

detectOS()
}
\value{
Character string: "windows", "macos", or "linux"

Character string: "windows", "macos", or "linux"
}
\description{
Simple OS detection utility

Detect the operating system
}
