% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.active_log_session}
\alias{.active_log_session}
\title{Get Active Log Session}
\usage{
.active_log_session()
}
\value{
Environment or `NULL`.
}
\description{
Returns the current buffered log session, if any.
}
\keyword{internal}
