% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.add_single_cells_cosmx}
\alias{.add_single_cells_cosmx}
\title{Add Single Cells Cosmx}
\usage{
.add_single_cells_cosmx(
  scope_obj,
  cosmx_root,
  filter_genes = NULL,
  exclude_prefix = c("Unassigned", "NegControl", "Background", "DeprecatedCodeword",
    "SystemControl", "Negative"),
  id_mode = c("cell_id", "fov_cell", "auto"),
  filter_barcodes = TRUE,
  force_all_genes = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{scope_obj}{A `scope_object`.}

\item{cosmx_root}{Parameter value.}

\item{filter_genes}{Parameter value.}

\item{exclude_prefix}{Parameter value.}

\item{id_mode}{Parameter value.}

\item{filter_barcodes}{Parameter value.}

\item{force_all_genes}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.add_single_cells_cosmx`.
}
\keyword{internal}
