% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.adjust_future_globals_maxsize}
\alias{.adjust_future_globals_maxsize}
\title{Adjust Future Globals Maxsize}
\usage{
.adjust_future_globals_maxsize(min_bytes = 2 * 1024^3)
}
\arguments{
\item{min_bytes}{Numeric threshold.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.adjust_future_globals_maxsize`.
}
\keyword{internal}
