% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.apply_ggfx_effect}
\alias{.apply_ggfx_effect}
\title{Apply ggfx effects (adapter)}
\usage{
.apply_ggfx_effect(plot, effect_type = "blend", ...)
}
\arguments{
\item{plot}{ggplot object}

\item{effect_type}{Type of effect ("blend", "reference", etc.)}

\item{...}{Additional arguments for the effect}
}
\value{
Modified ggplot object
}
\description{
Apply visual effects using ggfx
}
\keyword{internal}
