% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.as_numeric_or_na}
\alias{.as_numeric_or_na}
\title{As Numeric Or Na}
\usage{
.as_numeric_or_na(x)
}
\arguments{
\item{x}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.as_numeric_or_na`.
}
\keyword{internal}
