% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.as_sf}
\alias{.as_sf}
\title{Convert to sf object (adapter)}
\usage{
.as_sf(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Additional arguments}
}
\value{
sf object
}
\description{
Generic conversion to sf
}
\keyword{internal}
