% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.assemble_cosmx_sparse_matrix}
\alias{.assemble_cosmx_sparse_matrix}
\title{Assemble Cosmx Sparse Matrix}
\usage{
.assemble_cosmx_sparse_matrix(expr_dt, gene_cols, column_order)
}
\arguments{
\item{expr_dt}{Parameter value.}

\item{gene_cols}{Parameter value.}

\item{column_order}{Parameter value.}
}
\value{
A list with configuration/metadata used internally.
}
\description{
Internal helper for `.assemble_cosmx_sparse_matrix`.
}
\keyword{internal}
