% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.auto_histology_crop_bbox}
\alias{.auto_histology_crop_bbox}
\title{Auto Histology Crop Bbox}
\usage{
.auto_histology_crop_bbox(roi_bbox, microns_per_pixel, level_scalefactor)
}
\arguments{
\item{roi_bbox}{Parameter value.}

\item{microns_per_pixel}{Parameter value.}

\item{level_scalefactor}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.auto_histology_crop_bbox`.
}
\keyword{internal}
