% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.autoconfigure_networkit_runtime}
\alias{.autoconfigure_networkit_runtime}
\title{Autoconfigure Networkit Runtime}
\usage{
.autoconfigure_networkit_runtime()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.autoconfigure_networkit_runtime`.
}
\keyword{internal}
