% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.big_matrix_to_matrix}
\alias{.big_matrix_to_matrix}
\title{Convert big.matrix to regular matrix (adapter)}
\usage{
.big_matrix_to_matrix(bigmat)
}
\arguments{
\item{bigmat}{big.matrix object}
}
\value{
Regular R matrix
}
\description{
Wrapper around bigmemory::as.matrix
}
\keyword{internal}
