% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.bigmemory_available}
\alias{.bigmemory_available}
\title{Check if bigmemory is available}
\usage{
.bigmemory_available()
}
\value{
Logical
}
\description{
Check if bigmemory is available
}
\keyword{internal}
