% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.build_scope_from_visium_seurat}
\alias{.build_scope_from_visium_seurat}
\title{Build Scope From Visium Seurat}
\usage{
.build_scope_from_visium_seurat(
  input_dir,
  run_sctransform = TRUE,
  sct_return_only_var_genes = FALSE,
  vars_to_regress = NULL,
  glmGamPoi = TRUE,
  include_in_tissue_spots = TRUE,
  grid_multiplier = 1,
  flip_y = FALSE,
  verbose = TRUE,
  data_type = "visium"
)
}
\arguments{
\item{input_dir}{Filesystem path.}

\item{run_sctransform}{Parameter value.}

\item{sct_return_only_var_genes}{Parameter value.}

\item{vars_to_regress}{Parameter value.}

\item{glmGamPoi}{Parameter value.}

\item{include_in_tissue_spots}{Parameter value.}

\item{grid_multiplier}{Parameter value.}

\item{flip_y}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}

\item{data_type}{Parameter value.}
}
\value{
A list with configuration/metadata used internally.
}
\description{
Internal helper for `.build_scope_from_visium_seurat`.
}
\keyword{internal}
