% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.clip_points_to_polygon}
\alias{.clip_points_to_polygon}
\title{Clip Points To Polygon}
\usage{
.clip_points_to_polygon(points, polygon, chunk_size = 5e+05, ncores = 1)
}
\arguments{
\item{points}{Parameter value.}

\item{polygon}{Parameter value.}

\item{chunk_size}{Parameter value.}

\item{ncores}{Number of cores/threads to use.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.clip_points_to_polygon`.
}
\keyword{internal}
