% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.clip_points_within_roi}
\alias{.clip_points_within_roi}
\title{Clip Points Within Roi}
\usage{
.clip_points_within_roi(state)
}
\arguments{
\item{state}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.clip_points_within_roi`.
}
\keyword{internal}
