% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.cluster_metrics_sparse}
\alias{.cluster_metrics_sparse}
\title{Cluster Metrics Sparse}
\usage{
.cluster_metrics_sparse(members, cons_mat, W_mat, genes)
}
\arguments{
\item{members}{Parameter value.}

\item{cons_mat}{Parameter value.}

\item{W_mat}{Parameter value.}

\item{genes}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.cluster_metrics_sparse`.
}
\keyword{internal}
