% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.cluster_timestamp}
\alias{.cluster_timestamp}
\title{Cluster Timestamp}
\usage{
.cluster_timestamp()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.cluster_timestamp`.
}
\keyword{internal}
