% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.cmh_igraph_to_matrix}
\alias{.cmh_igraph_to_matrix}
\title{Cmh Igraph To Matrix}
\usage{
.cmh_igraph_to_matrix(g, genes = NULL, weight_cols = c("CMH", "weight"))
}
\arguments{
\item{g}{Parameter value.}

\item{genes}{Parameter value.}

\item{weight_cols}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.cmh_igraph_to_matrix`.
}
\keyword{internal}
