% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph.R
\name{.community_detect}
\alias{.community_detect}
\title{Community Detect}
\usage{
.community_detect(
  graph,
  algo = c("leiden", "louvain"),
  resolution = 1,
  objective = c("CPM", "modularity"),
  backend = c("igraph", "networkit"),
  threads = 1,
  debug = FALSE
)
}
\arguments{
\item{graph}{Parameter value.}

\item{algo}{Parameter value.}

\item{resolution}{Parameter value.}

\item{objective}{Parameter value.}

\item{backend}{Parameter value.}

\item{threads}{Number of threads to use.}

\item{debug}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.community_detect`.
}
\keyword{internal}
