% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph.R
\name{.community_detect_networkit}
\alias{.community_detect_networkit}
\title{Community Detect Networkit}
\usage{
.community_detect_networkit(graph, resolution = 1, threads = 1, debug = FALSE)
}
\arguments{
\item{graph}{Parameter value.}

\item{resolution}{Parameter value.}

\item{threads}{Number of threads to use.}

\item{debug}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.community_detect_networkit`.
}
\keyword{internal}
