% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.complexheatmap_available}
\alias{.complexheatmap_available}
\title{Check if ComplexHeatmap is available}
\usage{
.complexheatmap_available()
}
\value{
Logical
}
\description{
Check if ComplexHeatmap is available
}
\keyword{internal}
