% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.compute_auprc}
\alias{.compute_auprc}
\title{Compute AUPRC for a score ranking.}
\usage{
.compute_auprc(labels, scores)
}
\description{
Compute AUPRC for a score ranking.
}
\keyword{internal}
