% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.compute_auroc}
\alias{.compute_auroc}
\title{Compute AUROC for a score ranking.}
\usage{
.compute_auroc(labels, scores)
}
\description{
Compute AUROC for a score ranking.
}
\keyword{internal}
