% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_stats_correlation.R
\name{.compute_block_correlation}
\alias{.compute_block_correlation}
\title{Compute correlation between two sparse matrix blocks}
\usage{
.compute_block_correlation(block_i, block_j, means_i, means_j, sds_i, sds_j)
}
\arguments{
\item{block_i}{First block}

\item{block_j}{Second block}

\item{means_i}{Column means of first block}

\item{means_j}{Column means of second block}

\item{sds_i}{Column standard deviations of first block}

\item{sds_j}{Column standard deviations of second block}
}
\value{
Correlation matrix block
}
\description{
Internal helper for `.compute_block_correlation`.
}
\keyword{internal}
