% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.compute_l_assemble_outputs}
\alias{.compute_l_assemble_outputs}
\title{Internal helper for Lee's L workflows}
\usage{
.compute_l_assemble_outputs(
  scope_obj,
  gname,
  layer_name,
  L,
  Z_mat,
  P,
  betas,
  L_min,
  qc,
  FDR_main,
  FDR_storey,
  FDR_out_disc,
  FDR_out_beta,
  FDR_out_mid,
  FDR_out_uniform,
  perms,
  block_size,
  ncores,
  pi0_hat,
  n_sig_005,
  min_p_possible,
  FDR_main_method,
  input_cache,
  cache_inputs
)
}
\arguments{
\item{scope_obj}{Internal parameter}

\item{gname}{Parameter value.}

\item{layer_name}{Parameter value.}

\item{L}{Parameter value.}

\item{Z_mat}{Parameter value.}

\item{P}{Parameter value.}

\item{betas}{Parameter value.}

\item{L_min}{Numeric threshold.}

\item{qc}{Parameter value.}

\item{FDR_main}{Parameter value.}

\item{FDR_storey}{Parameter value.}

\item{FDR_out_disc}{Parameter value.}

\item{FDR_out_beta}{Parameter value.}

\item{FDR_out_mid}{Parameter value.}

\item{FDR_out_uniform}{Parameter value.}

\item{perms}{Parameter value.}

\item{block_size}{Parameter value.}

\item{ncores}{Number of cores/threads to use.}

\item{pi0_hat}{Parameter value.}

\item{n_sig_005}{Parameter value.}

\item{min_p_possible}{Numeric threshold.}

\item{FDR_main_method}{Parameter value.}

\item{input_cache}{Parameter value.}

\item{cache_inputs}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.compute_l_assemble_outputs`.
}
\keyword{internal}
