% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.compute_l_postprocess}
\alias{.compute_l_postprocess}
\title{Internal helper for Lee's L workflows}
\usage{
.compute_l_postprocess(
  L,
  X_full,
  X_used,
  W,
  grid_inf,
  cells,
  genes,
  within,
  chunk_size,
  L_min,
  block_id,
  perms,
  block_size,
  current_cores,
  verbose
)
}
\arguments{
\item{L}{Internal parameter}

\item{X_full}{Parameter value.}

\item{X_used}{Parameter value.}

\item{W}{Parameter value.}

\item{grid_inf}{Parameter value.}

\item{cells}{Parameter value.}

\item{genes}{Parameter value.}

\item{within}{Parameter value.}

\item{chunk_size}{Parameter value.}

\item{L_min}{Numeric threshold.}

\item{block_id}{Parameter value.}

\item{perms}{Parameter value.}

\item{block_size}{Parameter value.}

\item{current_cores}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.compute_l_postprocess`.
}
\keyword{internal}
