% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.compute_l_resolve_paths_and_cache}
\alias{.compute_l_resolve_paths_and_cache}
\title{Internal helper for Lee's L workflows}
\usage{
.compute_l_resolve_paths_and_cache(validation)
}
\arguments{
\item{validation}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.compute_l_resolve_paths_and_cache`.
}
\keyword{internal}
