% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.compute_mh_assemble_outputs}
\alias{.compute_mh_assemble_outputs}
\title{Internal helper for metrics/tests helpers}
\usage{
.compute_mh_assemble_outputs(
  scope_obj,
  gname,
  lee_layer,
  graph_slot_mh,
  matrix_slot_mh,
  gnet,
  MH,
  out,
  verbose
)
}
\arguments{
\item{scope_obj}{Internal parameter}

\item{gname}{Parameter value.}

\item{lee_layer}{Layer name.}

\item{graph_slot_mh}{Slot name.}

\item{matrix_slot_mh}{Slot name.}

\item{gnet}{Parameter value.}

\item{MH}{Parameter value.}

\item{out}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.compute_mh_assemble_outputs`.
}
\keyword{internal}
