% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.compute_mh_resolve_runtime_config}
\alias{.compute_mh_resolve_runtime_config}
\title{Internal helper for metrics/tests helpers}
\usage{
.compute_mh_resolve_runtime_config(ncores)
}
\arguments{
\item{ncores}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.compute_mh_resolve_runtime_config`.
}
\keyword{internal}
