% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_module_quality_metrics.R
\name{.compute_weighted_modularity}
\alias{.compute_weighted_modularity}
\title{Compute weighted modularity from a membership vector.}
\usage{
.compute_weighted_modularity(graph, membership, edge_weight = "weight")
}
\description{
Compute weighted modularity from a membership vector.
}
\keyword{internal}
