% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.Pearson.r
\name{.computeCorrelationBigMatrix}
\alias{.computeCorrelationBigMatrix}
\title{Use bigmemory for chunked correlation computation}
\usage{
.computeCorrelationBigMatrix(
  mat,
  method,
  chunk_size,
  ncores,
  force_compute,
  memory_limit_gb,
  backing_path,
  verbose = TRUE
)
}
\arguments{
\item{mat}{Expression matrix}

\item{method}{Correlation method}

\item{chunk_size}{Chunk size}

\item{ncores}{Number of threads}

\item{force_compute}{Whether to force computation}

\item{memory_limit_gb}{Memory limit}

\item{backing_path}{File storage path}

\item{verbose}{Logical. Whether to print progress messages (default TRUE).}
}
\value{
File-backed big.matrix object or error message
}
\description{
For extremely large matrices, use file-backed storage to avoid memory overflow
}
