% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.configure_io_bound}
\alias{.configure_io_bound}
\title{Configure for I/O bound operations}
\usage{
.configure_io_bound(ncores_requested)
}
\arguments{
\item{ncores_requested}{Parameter value.}
}
\value{
A list with configuration/metadata used internally.
}
\description{
Internal helper for `.configure_io_bound`.
}
\keyword{internal}
