% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.configure_open_mp_only}
\alias{.configure_open_mp_only}
\title{Configure for OpenMP-only operations}
\usage{
.configure_open_mp_only(ncores_requested)
}
\arguments{
\item{ncores_requested}{Parameter value.}
}
\value{
A list with configuration/metadata used internally.
}
\description{
Internal helper for `.configure_open_mp_only`.
}
\keyword{internal}
