% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_threading.r
\name{.configureComputeIntensive}
\alias{.configureComputeIntensive}
\title{Configure for compute-intensive operations}
\usage{
.configureComputeIntensive(ncores_requested)
}
\description{
Configure for compute-intensive operations
}
