% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_threading.r
\name{.configureOpenMPOnly}
\alias{.configureOpenMPOnly}
\title{Configure for OpenMP-only operations}
\usage{
.configureOpenMPOnly(ncores_requested)
}
\description{
Configure for OpenMP-only operations
}
