% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_modules_stage6.R
\name{.convert_flag}
\alias{.convert_flag}
\title{Convert Flag}
\usage{
.convert_flag(value)
}
\arguments{
\item{value}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.convert_flag`.
}
\keyword{internal}
