% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.create_advanced_heatmap}
\alias{.create_advanced_heatmap}
\title{Create advanced heatmap (adapter)}
\usage{
.create_advanced_heatmap(matrix, ...)
}
\arguments{
\item{matrix}{Matrix to plot}

\item{...}{Additional arguments passed to Heatmap}
}
\value{
Heatmap object
}
\description{
Wrapper around ComplexHeatmap::Heatmap
}
\keyword{internal}
