% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.create_big_matrix}
\alias{.create_big_matrix}
\title{Create big.matrix (adapter)}
\usage{
.create_big_matrix(nrow, ncol, type = "double", backingfile = NULL, ...)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{type}{Type of matrix (default "double")}

\item{backingfile}{Optional backing file path}

\item{...}{Additional arguments}
}
\value{
big.matrix object
}
\description{
Wrapper around bigmemory::big.matrix
}
\keyword{internal}
