% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.create_color_ramp}
\alias{.create_color_ramp}
\title{Create color ramp for heatmaps (adapter)}
\usage{
.create_color_ramp(breaks, colors)
}
\arguments{
\item{breaks}{Break points}

\item{colors}{Colors for each break}
}
\value{
Color mapping function
}
\description{
Wrapper around circlize::colorRamp2
}
\keyword{internal}
