% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.create_geometry_collection}
\alias{.create_geometry_collection}
\title{Create simple feature geometry collection (adapter)}
\usage{
.create_geometry_collection(..., crs = NA)
}
\arguments{
\item{...}{Geometry objects}

\item{crs}{Coordinate reference system}
}
\value{
sfc object
}
\description{
Wrapper around sf::st_sfc
}
\keyword{internal}
