% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.create_polygon}
\alias{.create_polygon}
\title{Create polygon geometry (adapter)}
\usage{
.create_polygon(x)
}
\arguments{
\item{x}{List of matrices defining polygon rings}
}
\value{
Polygon geometry
}
\description{
Wrapper around sf::st_polygon
}
\keyword{internal}
