% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.create_sf_object}
\alias{.create_sf_object}
\title{Create sf object (adapter)}
\usage{
.create_sf_object(x, ...)
}
\arguments{
\item{x}{Data frame or matrix with coordinates}

\item{...}{Additional arguments passed to st_as_sf}
}
\value{
sf object
}
\description{
Wrapper around sf::st_as_sf
}
\keyword{internal}
