% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.dendextend_available}
\alias{.dendextend_available}
\title{Check if dendextend is available}
\usage{
.dendextend_available()
}
\value{
Logical
}
\description{
Check if dendextend is available
}
\keyword{internal}
