% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.dendsort_available}
\alias{.dendsort_available}
\title{Check if dendsort is available}
\usage{
.dendsort_available()
}
\value{
Logical
}
\description{
Check if dendsort is available
}
\keyword{internal}
