% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.detect_container}
\alias{.detect_container}
\title{Detect Container}
\usage{
.detect_container()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.detect_container`.
}
\keyword{internal}
