% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.deterministic_arrow_threads}
\alias{.deterministic_arrow_threads}
\title{Deterministic Arrow Threads}
\usage{
.deterministic_arrow_threads(ncores_requested)
}
\arguments{
\item{ncores_requested}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.deterministic_arrow_threads`.
}
\keyword{internal}
