% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_matrix_utils.R
\name{.edges_from_upper_sparse}
\alias{.edges_from_upper_sparse}
\title{Upper-triangle edges from sparse matrix}
\usage{
.edges_from_upper_sparse(M, predicate = c("gt0", "ne0"), keep_diag = FALSE)
}
\arguments{
\item{M}{Matrix-like object.}

\item{predicate}{One of: "gt0", "ne0".}

\item{keep_diag}{Include diagonal.}
}
\value{
Integer matrix (i, j) in upper triangle.
}
\description{
Helper that returns upper-triangle edges from a matrix predicate.
}
\keyword{internal}
