% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.ensure_numeric_matrix}
\alias{.ensure_numeric_matrix}
\title{Ensure Numeric Matrix}
\usage{
.ensure_numeric_matrix(
  mat,
  nrow = NULL,
  ncol = NULL,
  rownames_out = NULL,
  colnames_out = NULL,
  as_sparse = FALSE
)
}
\arguments{
\item{mat}{Parameter value.}

\item{nrow}{Parameter value.}

\item{ncol}{Parameter value.}

\item{rownames_out}{Parameter value.}

\item{colnames_out}{Parameter value.}

\item{as_sparse}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.ensure_numeric_matrix`.
}
\keyword{internal}
