% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_ref_read}
\alias{.external_ref_read}
\title{Read external reference payload from scope_obj@stats.}
\usage{
.external_ref_read(
  scope_obj,
  grid_name,
  stats_layer,
  graph_slot,
  ref = "STRINGdb"
)
}
\description{
Read external reference payload from scope_obj@stats.
}
\keyword{internal}
