% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_metrics.R
\name{.external_reference_enrichment}
\alias{.external_reference_enrichment}
\title{Compute enrichment of STRING-positive edges vs background.}
\usage{
.external_reference_enrichment(n_pos_graph, n_graph, n_pos_bg, n_pairs_bg)
}
\description{
Compute enrichment of STRING-positive edges vs background.
}
\keyword{internal}
