% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_reference_hash}
\alias{.external_reference_hash}
\title{Hash helper for external reference caches.}
\usage{
.external_reference_hash(x)
}
\description{
Hash helper for external reference caches.
}
\keyword{internal}
