% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_reference_params_digest}
\alias{.external_reference_params_digest}
\title{Build digest of scoring parameters.}
\usage{
.external_reference_params_digest(params)
}
\description{
Build digest of scoring parameters.
}
\keyword{internal}
