% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_reference_pr_curve}
\alias{.external_reference_pr_curve}
\title{Compute PR curve points.}
\usage{
.external_reference_pr_curve(labels, scores)
}
\description{
Compute PR curve points.
}
\keyword{internal}
