% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_metrics.R
\name{.external_reference_precision_at_k}
\alias{.external_reference_precision_at_k}
\title{Compute precision at K for a score ranking.}
\usage{
.external_reference_precision_at_k(labels, scores, k = c(50L, 100L, 500L))
}
\description{
Compute precision at K for a score ranking.
}
\keyword{internal}
