% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_reference_roc_curve}
\alias{.external_reference_roc_curve}
\title{Compute ROC curve points.}
\usage{
.external_reference_roc_curve(labels, scores)
}
\description{
Compute ROC curve points.
}
\keyword{internal}
