% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_reference_top_enrichment}
\alias{.external_reference_top_enrichment}
\title{Compute enrichment for defined top-edge sets.}
\usage{
.external_reference_top_enrichment(
  edge_df,
  score_threshold,
  expected_rate,
  set_defs
)
}
\description{
Compute enrichment for defined top-edge sets.
}
\keyword{internal}
