% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.external_reference_top_sets}
\alias{.external_reference_top_sets}
\title{Define top-edge sets for enrichment evaluation.}
\usage{
.external_reference_top_sets(n_edges, precision_k, top_frac = NULL)
}
\description{
Define top-edge sets for enrichment evaluation.
}
\keyword{internal}
