% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_extractors.R
\name{.extract_edge_table_from_igraph}
\alias{.extract_edge_table_from_igraph}
\title{Extract standardized edge table from an igraph.}
\usage{
.extract_edge_table_from_igraph(graph, edge_weight = "weight")
}
\description{
Extract standardized edge table from an igraph.
}
\keyword{internal}
