% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_extractors.R
\name{.extract_graph_any}
\alias{.extract_graph_any}
\title{Extract igraph from supported inputs.}
\usage{
.extract_graph_any(
  x,
  grid_name = "grid30",
  stats_layer = "LeeStats_Xz",
  graph_slot = "g_consensus"
)
}
\description{
Extract igraph from supported inputs.
}
\keyword{internal}
