% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_extractors.R
\name{.extract_graph_from_dendro_return}
\alias{.extract_graph_from_dendro_return}
\title{Extract graph from plotDendroNetwork() output.}
\usage{
.extract_graph_from_dendro_return(x)
}
\description{
Extract graph from plotDendroNetwork() output.
}
\keyword{internal}
