% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_extractors.R
\name{.extract_graph_from_scope}
\alias{.extract_graph_from_scope}
\title{Extract graph from scope_object using stats path.}
\usage{
.extract_graph_from_scope(
  scope_obj,
  grid_name = "grid30",
  stats_layer = "LeeStats_Xz",
  graph_slot = "g_consensus"
)
}
\description{
Extract graph from scope_object using stats path.
}
\keyword{internal}
