% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_extractors.R
\name{.extract_membership_any}
\alias{.extract_membership_any}
\title{Extract membership vector from supported inputs.}
\usage{
.extract_membership_any(
  x,
  graph,
  membership = NULL,
  grid_name = "grid30",
  stats_layer = "LeeStats_Xz",
  graph_slot = "g_consensus"
)
}
\description{
Extract membership vector from supported inputs.
}
\keyword{internal}
