% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.filter_matrix_by_quantile}
\alias{.filter_matrix_by_quantile}
\title{Filter Matrix By Quantile}
\usage{
.filter_matrix_by_quantile(mat, pct_min = "q0", pct_max = "q100")
}
\arguments{
\item{mat}{Parameter value.}

\item{pct_min}{Numeric threshold.}

\item{pct_max}{Numeric threshold.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.filter_matrix_by_quantile`.
}
\keyword{internal}
