% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.fix_rng}
\alias{.fix_rng}
\title{Fix Rng}
\usage{
.fix_rng(seed = 42)
}
\arguments{
\item{seed}{Random seed.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.fix_rng`.
}
\keyword{internal}
