% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.flip_coordinates}
\alias{.flip_coordinates}
\title{Flip Coordinates}
\usage{
.flip_coordinates(data, y_max)
}
\arguments{
\item{data}{Parameter value.}

\item{y_max}{Numeric threshold.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.flip_coordinates`.
}
\keyword{internal}
