% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.flip_y_coordinates}
\alias{.flip_y_coordinates}
\title{Flip Y Coordinates}
\usage{
.flip_y_coordinates(coords, y_max)
}
\arguments{
\item{coords}{Parameter value.}

\item{y_max}{Numeric threshold.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.flip_y_coordinates`.
}
\keyword{internal}
