% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.force_blas_single_thread}
\alias{.force_blas_single_thread}
\title{Force Blas Single Thread}
\usage{
.force_blas_single_thread()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.force_blas_single_thread`.
}
\keyword{internal}
