% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.force_single_thread_blas}
\alias{.force_single_thread_blas}
\title{Force Single Thread Blas}
\usage{
.force_single_thread_blas()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.force_single_thread_blas`.
}
\keyword{internal}
