% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_threading.r
\name{.forceBLASSingleThread}
\alias{.forceBLASSingleThread}
\title{Force BLAS to single thread}
\usage{
.forceBLASSingleThread()
}
\description{
Aggressively sets all known BLAS libraries to single thread
}
