% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.format_label_display}
\alias{.format_label_display}
\title{Format Label Display}
\usage{
.format_label_display(x)
}
\arguments{
\item{x}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.format_label_display`.
}
\keyword{internal}
