% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_parallel.R
\name{.future_apply_available}
\alias{.future_apply_available}
\title{Check if future.apply is available}
\usage{
.future_apply_available()
}
\value{
Logical
}
\description{
Check if future.apply is available
}
\keyword{internal}
