% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.future_maxsize_env_override}
\alias{.future_maxsize_env_override}
\title{Future Maxsize Env Override}
\usage{
.future_maxsize_env_override()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.future_maxsize_env_override`.
}
\keyword{internal}
