% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.gene_metrics}
\alias{.gene_metrics}
\title{Gene Metrics}
\usage{
.gene_metrics(members, cons_mat, W_mat)
}
\arguments{
\item{members}{Parameter value.}

\item{cons_mat}{Parameter value.}

\item{W_mat}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.gene_metrics`.
}
\keyword{internal}
