% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.get_bbox}
\alias{.get_bbox}
\title{Get bounding box (adapter)}
\usage{
.get_bbox(obj)
}
\arguments{
\item{obj}{sf object or extent}
}
\value{
Named numeric vector with bounding box
}
\description{
Wrapper around sf::st_bbox
}
\keyword{internal}
