% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.get_geometry}
\alias{.get_geometry}
\title{Get geometry from sf object (adapter)}
\usage{
.get_geometry(x)
}
\arguments{
\item{x}{sf object}
}
\value{
sfc geometry list column
}
\description{
Wrapper around sf::st_geometry
}
\keyword{internal}
