% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.get_or_set_crs}
\alias{.get_or_set_crs}
\title{Get/set coordinate reference system (adapter)}
\usage{
.get_or_set_crs(x, value = NULL)
}
\arguments{
\item{x}{sf object}

\item{value}{CRS value to set (optional)}
}
\value{
CRS object or modified sf object
}
\description{
Wrapper around sf::st_crs
}
\keyword{internal}
