% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_stats_correlation.R
\name{.get_pearson_matrix}
\alias{.get_pearson_matrix}
\title{Get Pearson Matrix}
\usage{
.get_pearson_matrix(scope_obj, grid_name = NULL, level = c("grid", "cell"))
}
\arguments{
\item{scope_obj}{A `scope_object`.}

\item{grid_name}{Grid layer name (or `NULL` to use the active layer).}

\item{level}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.get_pearson_matrix`.
}
\keyword{internal}
