% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_visium.R
\name{.get_seurat_assay_data}
\alias{.get_seurat_assay_data}
\title{Get assay data from Seurat object (adapter)}
\usage{
.get_seurat_assay_data(seurat_obj, assay = "SCT", slot = "data")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{assay}{Assay name (default "SCT")}

\item{slot}{Slot name (default "data")}
}
\value{
Matrix of assay data
}
\description{
Wrapper around SeuratObject::GetAssayData
}
\keyword{internal}
