% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_external_reference.R
\name{.get_string_interactions}
\alias{.get_string_interactions}
\title{Get STRING interactions (adapter)}
\usage{
.get_string_interactions(string_db, string_ids)
}
\arguments{
\item{string_db}{STRINGdb object}

\item{string_ids}{Character vector of STRING protein IDs}
}
\value{
Data frame with interactions (from, to, combined_score)
}
\description{
Retrieve protein-protein interactions from STRING
}
\keyword{internal}
