% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.get_system_memory_gb}
\alias{.get_system_memory_gb}
\title{Get System Memory Gb}
\usage{
.get_system_memory_gb()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.get_system_memory_gb`.
}
\keyword{internal}
