% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.get_top_lvs_r_assemble_outputs}
\alias{.get_top_lvs_r_assemble_outputs}
\title{Internal helper for Lee's L workflows}
\usage{
.get_top_lvs_r_assemble_outputs(sel)
}
\arguments{
\item{sel}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.get_top_lvs_r_assemble_outputs`.
}
\keyword{internal}
