% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.get_top_lvs_r_resolve_runtime_config}
\alias{.get_top_lvs_r_resolve_runtime_config}
\title{Internal helper for Lee's L workflows}
\usage{
.get_top_lvs_r_resolve_runtime_config(ncores, verbose)
}
\arguments{
\item{ncores}{Internal parameter}

\item{verbose}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.get_top_lvs_r_resolve_runtime_config`.
}
\keyword{internal}
