% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_visium.R
\name{.glmgampoi_available}
\alias{.glmgampoi_available}
\title{Check if glmGamPoi is available (for SCTransform)}
\usage{
.glmgampoi_available()
}
\value{
Logical indicating if glmGamPoi is installed
}
\description{
Check if glmGamPoi is available (for SCTransform)
}
\keyword{internal}
