% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_module_quality_metrics.R
\name{.graph_edge_table}
\alias{.graph_edge_table}
\title{Build edge table with score column from an igraph.}
\usage{
.graph_edge_table(graph, edge_weight = "weight")
}
\description{
Build edge table with score column from an igraph.
}
\keyword{internal}
