% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_spatial_weights.R
\name{.grid_nb_hex_r}
\alias{.grid_nb_hex_r}
\title{Hex grid R implementation (odd-r/even-r)}
\usage{
.grid_nb_hex_r(nrow, ncol, oddr = TRUE)
}
\description{
Hex grid R implementation (odd-r/even-r)
}
\keyword{internal}
