% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_spatial_weights.R
\name{.grid_nb_hexq_r}
\alias{.grid_nb_hexq_r}
\title{Hex grid R implementation (odd-q/even-q)}
\usage{
.grid_nb_hexq_r(nrow, ncol, oddq = TRUE)
}
\description{
Hex grid R implementation (odd-q/even-q)
}
\keyword{internal}
