% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.guess_singlecell_platform}
\alias{.guess_singlecell_platform}
\title{Guess Singlecell Platform}
\usage{
.guess_singlecell_platform(path)
}
\arguments{
\item{path}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.guess_singlecell_platform`.
}
\keyword{internal}
