% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.identify_cosmx_gene_columns}
\alias{.identify_cosmx_gene_columns}
\title{Identify Cosmx Gene Columns}
\usage{
.identify_cosmx_gene_columns(cols, exclude_prefix)
}
\arguments{
\item{cols}{Parameter value.}

\item{exclude_prefix}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.identify_cosmx_gene_columns`.
}
\keyword{internal}
