% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.infer_visium_mpp}
\alias{.infer_visium_mpp}
\title{Infer Visium Mpp}
\usage{
.infer_visium_mpp(
  grid_layer,
  scalefactors,
  default_spot_um = getOption("geneSCOPE.default_visium_spot_um", 55)
)
}
\arguments{
\item{grid_layer}{Layer name.}

\item{scalefactors}{Parameter value.}

\item{default_spot_um}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.infer_visium_mpp`.
}
\keyword{internal}
