% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.ingest_segmentation_geometries}
\alias{.ingest_segmentation_geometries}
\title{Ingest Segmentation Geometries}
\usage{
.ingest_segmentation_geometries(state)
}
\arguments{
\item{state}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.ingest_segmentation_geometries`.
}
\keyword{internal}
