% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.init_thread_manager}
\alias{.init_thread_manager}
\title{Init Thread Manager}
\usage{
.init_thread_manager()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.init_thread_manager`.
}
\keyword{internal}
