% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.initialize_output_object}
\alias{.initialize_output_object}
\title{Initialize Output Object}
\usage{
.initialize_output_object(state)
}
\arguments{
\item{state}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.initialize_output_object`.
}
\keyword{internal}
