% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_parallel.R
\name{.is_parallel_mode}
\alias{.is_parallel_mode}
\title{Check if running in parallel mode}
\usage{
.is_parallel_mode()
}
\value{
Logical
}
\description{
Check if future is configured for parallel execution
}
\keyword{internal}
