% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_leesL_compare.R
\name{.is_scope_object}
\alias{.is_scope_object}
\title{Internal helpers for cross-sample Lee's L comparisons}
\usage{
.is_scope_object(x)
}
\description{
Internal helpers for cross-sample Lee's L comparisons
}
\keyword{internal}
