% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_leesL.R
\name{.lee_l_full}
\alias{.lee_l_full}
\title{Lee L Full}
\usage{
.lee_l_full(Xz, W, n_threads = 1L, .cache = TRUE)
}
\arguments{
\item{Xz}{Parameter value.}

\item{W}{Parameter value.}

\item{n_threads}{Number of threads to use.}

\item{.cache}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.lee_l_full`.
}
\keyword{internal}
