% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_leesL.R
\name{.lee_l_subset}
\alias{.lee_l_subset}
\title{Lee L Subset}
\usage{
.lee_l_subset(Xz, W, cols0, n_threads = 1L)
}
\arguments{
\item{Xz}{Parameter value.}

\item{W}{Parameter value.}

\item{cols0}{Parameter value.}

\item{n_threads}{Number of threads to use.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.lee_l_subset`.
}
\keyword{internal}
