% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.limit_thread_environment}
\alias{.limit_thread_environment}
\title{Limit Thread Environment}
\usage{
.limit_thread_environment(state)
}
\arguments{
\item{state}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.limit_thread_environment`.
}
\keyword{internal}
