% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_visium.R
\name{.load_visium_data}
\alias{.load_visium_data}
\title{Load Visium data using Seurat (adapter)}
\usage{
.load_visium_data(data_dir, ...)
}
\arguments{
\item{data_dir}{Path to Visium data directory}

\item{...}{Additional arguments passed to Seurat::Load10X_Spatial}
}
\value{
Seurat object
}
\description{
Wrapper around Seurat::Load10X_Spatial with dependency checking
}
\keyword{internal}
