% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.locate_cosmx_dataset}
\alias{.locate_cosmx_dataset}
\title{Locate the CosMx dataset and expression matrix.}
\usage{
.locate_cosmx_dataset(cosmx_root)
}
\arguments{
\item{cosmx_root}{Root directory of the CosMx project.}
}
\value{
List containing dataset directory and expression matrix path.
}
\description{
Internal helper for `.locate_cosmx_dataset`.
}
\keyword{internal}
