% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.log_error}
\alias{.log_error}
\title{Log Error}
\usage{
.log_error(parent, msg)
}
\arguments{
\item{parent}{Parent function name.}

\item{msg}{Error message.}
}
\value{
Invisible NULL.
}
\description{
Convenience wrapper for emitting ERROR level logs.
}
\keyword{internal}
