% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.log_prefix}
\alias{.log_prefix}
\title{Logging Prefix}
\usage{
.log_prefix(parent, step)
}
\arguments{
\item{parent}{Parent function name.}

\item{step}{Step id.}
}
\value{
String prefix for structured logs.
}
\description{
Internal helper for .log_prefix.
}
\keyword{internal}
