% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_runtime_utils.R
\name{.log_warn}
\alias{.log_warn}
\title{Log Warning}
\usage{
.log_warn(parent, msg)
}
\arguments{
\item{parent}{Parent function name.}

\item{msg}{Warning message.}
}
\value{
Invisible NULL.
}
\description{
Convenience wrapper for emitting WARN level logs.
}
\keyword{internal}
