% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.lvsr_curve_compute}
\alias{.lvsr_curve_compute}
\title{Internal helper for Lee's L workflows}
\usage{
.lvsr_curve_compute(
  Lv,
  rv,
  downsample,
  jitter_eps,
  n_strata,
  length_out,
  span,
  deg,
  B,
  k_max,
  min_rel_width,
  widen_span,
  ci_method,
  ci_adjust,
  ncores,
  verbose
)
}
\arguments{
\item{Lv}{Internal parameter}

\item{rv}{Parameter value.}

\item{downsample}{Parameter value.}

\item{jitter_eps}{Parameter value.}

\item{n_strata}{Parameter value.}

\item{length_out}{Parameter value.}

\item{span}{Parameter value.}

\item{deg}{Parameter value.}

\item{B}{Parameter value.}

\item{k_max}{Numeric threshold.}

\item{min_rel_width}{Numeric threshold.}

\item{widen_span}{Parameter value.}

\item{ci_method}{Parameter value.}

\item{ci_adjust}{Parameter value.}

\item{ncores}{Number of cores/threads to use.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.lvsr_curve_compute`.
}
\keyword{internal}
