% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.lvsr_postprocess_and_return}
\alias{.lvsr_postprocess_and_return}
\title{Internal helper for Lee's L workflows}
\usage{
.lvsr_postprocess_and_return(
  sel,
  do_perm,
  perms,
  p_adj_mode,
  pval_mode,
  total_universe,
  scope_obj,
  grid_name,
  lee_stats_layer,
  expr_layer,
  use_blocks,
  block_side,
  mem_limit_GB,
  ncores,
  clamp_mode,
  verbose
)
}
\arguments{
\item{sel}{Internal parameter}

\item{do_perm}{Parameter value.}

\item{perms}{Parameter value.}

\item{p_adj_mode}{Parameter value.}

\item{pval_mode}{Parameter value.}

\item{total_universe}{Parameter value.}

\item{scope_obj}{A `scope_object`.}

\item{grid_name}{Grid layer name (or `NULL` to use the active layer).}

\item{lee_stats_layer}{Layer name.}

\item{expr_layer}{Layer name.}

\item{use_blocks}{Logical flag.}

\item{block_side}{Parameter value.}

\item{mem_limit_GB}{Parameter value.}

\item{ncores}{Number of cores/threads to use.}

\item{clamp_mode}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.lvsr_postprocess_and_return`.
}
\keyword{internal}
