% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.lvsr_prepare_lvs_r_table}
\alias{.lvsr_prepare_lvs_r_table}
\title{Internal helper for Lee's L workflows}
\usage{
.lvsr_prepare_lvs_r_table(
  matrices,
  scope_obj,
  grid_name,
  clamp_mode,
  direction,
  top_n,
  pear_range,
  L_range,
  curve_layer,
  CI_rule,
  lee_stats_layer,
  pear_level,
  verbose
)
}
\arguments{
\item{matrices}{Internal parameter}

\item{scope_obj}{A `scope_object`.}

\item{grid_name}{Grid layer name (or `NULL` to use the active layer).}

\item{clamp_mode}{Parameter value.}

\item{direction}{Parameter value.}

\item{top_n}{Parameter value.}

\item{pear_range}{Parameter value.}

\item{L_range}{Parameter value.}

\item{curve_layer}{Layer name.}

\item{CI_rule}{Parameter value.}

\item{lee_stats_layer}{Layer name.}

\item{pear_level}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.lvsr_prepare_lvs_r_table`.
}
\keyword{internal}
