% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.make_geometry_valid}
\alias{.make_geometry_valid}
\title{Make geometry valid (adapter)}
\usage{
.make_geometry_valid(x)
}
\arguments{
\item{x}{sf object}
}
\value{
sf object with valid geometries
}
\description{
Wrapper around sf::st_make_valid
}
\keyword{internal}
