% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.make_pal_map}
\alias{.make_pal_map}
\title{Make Pal Map}
\usage{
.make_pal_map(levels_vec, cluster_palette, tip_palette)
}
\arguments{
\item{levels_vec}{Parameter value.}

\item{cluster_palette}{Parameter value.}

\item{tip_palette}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.make_pal_map`.
}
\keyword{internal}
