% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph_core.R
\name{.networkit_edge_reader}
\alias{.networkit_edge_reader}
\title{Networkit Edge Reader}
\usage{
.networkit_edge_reader(nk)
}
\arguments{
\item{nk}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.networkit_edge_reader`.
}
\keyword{internal}
