% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph_core.R
\name{.networkit_plm_runs}
\alias{.networkit_plm_runs}
\title{Networkit Plm Runs}
\usage{
.networkit_plm_runs(
  vertex_names,
  edge_table,
  edge_weights,
  gamma = 1,
  refine = TRUE,
  threads = 1L,
  seeds = 1L
)
}
\arguments{
\item{vertex_names}{Parameter value.}

\item{edge_table}{Parameter value.}

\item{edge_weights}{Parameter value.}

\item{gamma}{Parameter value.}

\item{refine}{Parameter value.}

\item{threads}{Number of threads to use.}

\item{seeds}{Integer vector of random seeds.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.networkit_plm_runs`.
}
\keyword{internal}
