% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph_core.R
\name{.networkit_write_edge_file}
\alias{.networkit_write_edge_file}
\title{Networkit Write Edge File}
\usage{
.networkit_write_edge_file(
  vertex_names,
  edge_table,
  edge_weights,
  tmpdir = NULL
)
}
\arguments{
\item{vertex_names}{Parameter value.}

\item{edge_table}{Parameter value.}

\item{edge_weights}{Parameter value.}

\item{tmpdir}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.networkit_write_edge_file`.
}
\keyword{internal}
