% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph.R
\name{.nk_available}
\alias{.nk_available}
\title{Check if the Python 'networkit' module is available through reticulate.}
\usage{
.nk_available()
}
\value{
Logical flag indicating module availability.
}
\description{
Internal helper for `.nk_available`.
}
\keyword{internal}
