% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.normalize_histology_y_origin}
\alias{.normalize_histology_y_origin}
\title{Normalize Histology Y Origin}
\usage{
.normalize_histology_y_origin(y_origin_choice)
}
\arguments{
\item{y_origin_choice}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.normalize_histology_y_origin`.
}
\keyword{internal}
