% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_extractors.R
\name{.normalize_input_to_igraph}
\alias{.normalize_input_to_igraph}
\title{Normalize inputs to an igraph.}
\usage{
.normalize_input_to_igraph(
  x,
  grid_name = "grid30",
  stats_layer = "LeeStats_Xz",
  graph_slot = "g_consensus"
)
}
\description{
Normalize inputs to an igraph.
}
\keyword{internal}
