% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_hooks.R
\name{.on_attach}
\alias{.on_attach}
\title{On Attach}
\usage{
.on_attach(libname, pkgname)
}
\arguments{
\item{libname}{Parameter value.}

\item{pkgname}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.on_attach`.
}
\keyword{internal}
