% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_hooks.R
\name{.on_load}
\alias{.on_load}
\title{On Load}
\usage{
.on_load(libname, pkgname)
}
\arguments{
\item{libname}{Parameter value.}

\item{pkgname}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.on_load`.
}
\keyword{internal}
