% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_hooks.R
\name{.on_unload}
\alias{.on_unload}
\title{On Unload}
\usage{
.on_unload(libpath)
}
\arguments{
\item{libpath}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.on_unload`.
}
\keyword{internal}
