% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.order_levels_numeric}
\alias{.order_levels_numeric}
\title{Order Levels Numeric}
\usage{
.order_levels_numeric(lev)
}
\arguments{
\item{lev}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.order_levels_numeric`.
}
\keyword{internal}
