% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_parallel.R
\name{.parallel_lapply}
\alias{.parallel_lapply}
\title{Parallel lapply (adapter)}
\usage{
.parallel_lapply(X, FUN, ..., future.scheduling = 1)
}
\arguments{
\item{X}{List or vector}

\item{FUN}{Function to apply}

\item{...}{Additional arguments passed to FUN}

\item{future.scheduling}{Scheduling strategy}
}
\value{
List of results
}
\description{
Wrapper around future.apply::future_lapply
}
\keyword{internal}
