% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_parallel.R
\name{.parallel_sapply}
\alias{.parallel_sapply}
\title{Parallel sapply (adapter)}
\usage{
.parallel_sapply(X, FUN, ..., simplify = TRUE, future.scheduling = 1)
}
\arguments{
\item{X}{List or vector}

\item{FUN}{Function to apply}

\item{...}{Additional arguments passed to FUN}

\item{simplify}{Simplify result (default TRUE)}

\item{future.scheduling}{Scheduling strategy}
}
\value{
Vector or matrix of results
}
\description{
Wrapper around future.apply::future_sapply
}
\keyword{internal}
