% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_add_annotations}
\alias{.plot_dendro_add_annotations}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_add_annotations(dend_info, tree_data, params)
}
\arguments{
\item{dend_info}{Internal parameter}

\item{tree_data}{Internal parameter}

\item{params}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_add_annotations`.
}
\keyword{internal}
