% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_build_dendrogram}
\alias{.plot_dendro_build_dendrogram}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_build_dendrogram(dist_data, params)
}
\arguments{
\item{dist_data}{Internal parameter}

\item{params}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_build_dendrogram`.
}
\keyword{internal}
