% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_finalize}
\alias{.plot_dendro_finalize}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_finalize(annotated, params)
}
\arguments{
\item{annotated}{Internal parameter}

\item{params}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_finalize`.
}
\keyword{internal}
