% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_network_branches_output}
\alias{.plot_dendro_network_branches_output}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_network_branches_output(plot_obj, prep)
}
\arguments{
\item{plot_obj}{Internal parameter}

\item{prep}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_network_branches_output`.
}
\keyword{internal}
