% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_network_branches_prepare}
\alias{.plot_dendro_network_branches_prepare}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_network_branches_prepare(...)
}
\arguments{
\item{...}{Additional arguments (currently unused).}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_network_branches_prepare`.
}
\keyword{internal}
