% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_network_with_branches_build_annotation_layers}
\alias{.plot_dendro_network_with_branches_build_annotation_layers}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_network_with_branches_build_annotation_layers(
  plot_obj,
  title = NULL,
  caption = NULL,
  legend_order = 1
)
}
\arguments{
\item{plot_obj}{Parameter value.}

\item{title}{Parameter value.}

\item{caption}{Parameter value.}

\item{legend_order}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_network_with_branches_build_annotation_layers`.
}
\keyword{internal}
