% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_network_with_branches_build_subcluster_colors}
\alias{.plot_dendro_network_with_branches_build_subcluster_colors}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_network_with_branches_build_subcluster_colors(
  clu,
  sub_attr,
  cluster_palette,
  subbranch_palette,
  method_final
)
}
\arguments{
\item{clu}{Parameter value.}

\item{sub_attr}{Parameter value.}

\item{cluster_palette}{Parameter value.}

\item{subbranch_palette}{Parameter value.}

\item{method_final}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_network_with_branches_build_subcluster_colors`.
}
\keyword{internal}
