% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_network_with_branches_dispatch_plots}
\alias{.plot_dendro_network_with_branches_dispatch_plots}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_network_with_branches_dispatch_plots(
  base_args,
  method_final,
  title,
  cluster_vec_base,
  cluster_vec_sub = NULL
)
}
\arguments{
\item{base_args}{Parameter value.}

\item{method_final}{Parameter value.}

\item{title}{Parameter value.}

\item{cluster_vec_base}{Parameter value.}

\item{cluster_vec_sub}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_network_with_branches_dispatch_plots`.
}
\keyword{internal}
