% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_dendro_prepare_tree_data}
\alias{.plot_dendro_prepare_tree_data}
\title{Internal helper for plotting workflows}
\usage{
.plot_dendro_prepare_tree_data(params)
}
\arguments{
\item{params}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_dendro_prepare_tree_data`.
}
\keyword{internal}
