% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_add_scale_bar}
\alias{.plot_density_add_scale_bar}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_add_scale_bar(
  p,
  prep,
  x_rng,
  y_rng,
  scale_bar_pos,
  scale_bar_corner,
  bar_len,
  arrow_pt,
  scale_bar_colour,
  scale_text_size,
  bar_offset,
  scale_bar_show
)
}
\arguments{
\item{p}{Internal parameter}

\item{prep}{Parameter value.}

\item{x_rng}{Parameter value.}

\item{y_rng}{Parameter value.}

\item{scale_bar_pos}{Parameter value.}

\item{scale_bar_corner}{Parameter value.}

\item{bar_len}{Parameter value.}

\item{arrow_pt}{Parameter value.}

\item{scale_bar_colour}{Parameter value.}

\item{scale_text_size}{Parameter value.}

\item{bar_offset}{Parameter value.}

\item{scale_bar_show}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_add_scale_bar`.
}
\keyword{internal}
