% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_apply_theme_axes}
\alias{.plot_density_apply_theme_axes}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_apply_theme_axes(p, prep, grid_gap, target_aspect_ratio)
}
\arguments{
\item{p}{Internal parameter}

\item{prep}{Internal parameter}

\item{grid_gap}{Internal parameter}

\item{target_aspect_ratio}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_apply_theme_axes`.
}
\keyword{internal}
