% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_build_layers}
\alias{.plot_density_build_layers}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_build_layers(
  prep,
  palette1,
  palette2,
  alpha1,
  alpha2,
  density1_name,
  density2_name,
  overlay_image,
  image_path,
  image_alpha,
  image_choice
)
}
\arguments{
\item{prep}{Internal parameter}

\item{palette1}{Parameter value.}

\item{palette2}{Parameter value.}

\item{alpha1}{Parameter value.}

\item{alpha2}{Parameter value.}

\item{density1_name}{Parameter value.}

\item{density2_name}{Parameter value.}

\item{overlay_image}{Parameter value.}

\item{image_path}{Filesystem path.}

\item{image_alpha}{Parameter value.}

\item{image_choice}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_build_layers`.
}
\keyword{internal}
