% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_build_overlays}
\alias{.plot_density_build_overlays}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_build_overlays(
  p,
  prep,
  seg_type,
  colour_cell,
  colour_nucleus,
  alpha_seg
)
}
\arguments{
\item{p}{Internal parameter}

\item{prep}{Parameter value.}

\item{seg_type}{Parameter value.}

\item{colour_cell}{Parameter value.}

\item{colour_nucleus}{Parameter value.}

\item{alpha_seg}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_build_overlays`.
}
\keyword{internal}
