% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_build_shape_geom}
\alias{.plot_density_build_shape_geom}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_build_shape_geom(df, alpha_val, tile_shape, hex_orientation)
}
\arguments{
\item{df}{Internal parameter}

\item{alpha_val}{Internal parameter}

\item{tile_shape}{Internal parameter}

\item{hex_orientation}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_build_shape_geom`.
}
\keyword{internal}
