% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_centroids_core}
\alias{.plot_density_centroids_core}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_centroids_core(
  scope_obj,
  gene1_name,
  gene2_name,
  palette1 = "#fc3d5d",
  palette2 = "#4753f8",
  size1 = 0.3,
  size2 = 0.3,
  alpha1 = 1,
  alpha2 = 0.5,
  seg_type = c("none", "cell", "nucleus", "both"),
  colour_cell = "black",
  colour_nucleus = "#3182bd",
  alpha_seg = 0.2,
  grid_gap = 100,
  scale_text_size = 2.4,
  bar_len = 400,
  bar_offset = 0.01,
  arrow_pt = 4,
  scale_legend_colour = "black",
  max.cutoff1 = 1,
  max.cutoff2 = 1
)
}
\arguments{
\item{scope_obj}{Internal parameter}

\item{gene1_name}{Parameter value.}

\item{gene2_name}{Parameter value.}

\item{palette1}{Parameter value.}

\item{palette2}{Parameter value.}

\item{size1}{Parameter value.}

\item{size2}{Parameter value.}

\item{alpha1}{Parameter value.}

\item{alpha2}{Parameter value.}

\item{seg_type}{Parameter value.}

\item{colour_cell}{Parameter value.}

\item{colour_nucleus}{Parameter value.}

\item{alpha_seg}{Parameter value.}

\item{grid_gap}{Parameter value.}

\item{scale_text_size}{Parameter value.}

\item{bar_len}{Parameter value.}

\item{bar_offset}{Parameter value.}

\item{arrow_pt}{Parameter value.}

\item{scale_legend_colour}{Parameter value.}

\item{max.cutoff1}{Parameter value.}

\item{max.cutoff2}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_centroids_core`.
}
\keyword{internal}
