% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_finalize}
\alias{.plot_density_finalize}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_finalize(p)
}
\arguments{
\item{p}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_finalize`.
}
\keyword{internal}
