% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_flip_raster}
\alias{.plot_density_flip_raster}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_flip_raster(img)
}
\arguments{
\item{img}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_flip_raster`.
}
\keyword{internal}
