% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_mirror}
\alias{.plot_density_mirror}
\title{Plot Density Mirror}
\usage{
.plot_density_mirror(
  scope_obj,
  layer_name,
  gene1,
  gene2,
  sort_by = c("sum", "diff", "gene1", "gene2"),
  rescale = TRUE
)
}
\arguments{
\item{scope_obj}{A `scope_object`.}

\item{layer_name}{Parameter value.}

\item{gene1}{Parameter value.}

\item{gene2}{Parameter value.}

\item{sort_by}{Parameter value.}

\item{rescale}{Parameter value.}
}
\value{
A `ggplot` object (or a list of plot components) used internally.
}
\description{
Internal helper for `.plot_density_mirror`.
}
\keyword{internal}
