% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_prepare_rgba}
\alias{.plot_density_prepare_rgba}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_prepare_rgba(img, alpha_scale = 1)
}
\arguments{
\item{img}{Internal parameter}

\item{alpha_scale}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_prepare_rgba`.
}
\keyword{internal}
