% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_resolve_scale_bar_pos}
\alias{.plot_density_resolve_scale_bar_pos}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_resolve_scale_bar_pos(
  pos,
  corner,
  default_x = 0.03,
  default_y = 0.22
)
}
\arguments{
\item{pos}{Internal parameter}

\item{corner}{Internal parameter}

\item{default_x}{Internal parameter}

\item{default_y}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_resolve_scale_bar_pos`.
}
\keyword{internal}
