% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_tile_df}
\alias{.plot_density_tile_df}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_tile_df(df)
}
\arguments{
\item{df}{Internal parameter}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_tile_df`.
}
\keyword{internal}
