% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_density_validate_inputs}
\alias{.plot_density_validate_inputs}
\title{Internal helper for plotting workflows}
\usage{
.plot_density_validate_inputs(
  tile_shape = c("square", "circle", "hex"),
  hex_orientation = c("flat", "pointy"),
  scale_bar_corner = c("bottom-left", "bottom-right", "top-left", "top-right"),
  seg_type = c("cell", "nucleus", "both"),
  image_choice = c("auto", "hires", "lowres"),
  histology_level = c("lowres", "hires"),
  axis_mode = c("grid", "image"),
  aspect_ratio = NULL
)
}
\arguments{
\item{tile_shape}{Internal parameter}

\item{hex_orientation}{Parameter value.}

\item{scale_bar_corner}{Parameter value.}

\item{seg_type}{Parameter value.}

\item{image_choice}{Parameter value.}

\item{histology_level}{Parameter value.}

\item{axis_mode}{Parameter value.}

\item{aspect_ratio}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_density_validate_inputs`.
}
\keyword{internal}
