% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_lvs_r_core}
\alias{.plot_lvs_r_core}
\title{Internal helper for plotting workflows}
\usage{
.plot_lvs_r_core(
  scope_obj,
  grid_name,
  pear_level = c("cell", "grid"),
  lee_stats_layer = "LeeStats_Xz",
  delta_top_n = 10,
  flip = TRUE
)
}
\arguments{
\item{scope_obj}{Internal parameter}

\item{grid_name}{Grid layer name (or `NULL` to use the active layer).}

\item{pear_level}{Parameter value.}

\item{lee_stats_layer}{Layer name.}

\item{delta_top_n}{Parameter value.}

\item{flip}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_lvs_r_core`.
}
\keyword{internal}
