% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_network_build_ggraph_layers}
\alias{.plot_network_build_ggraph_layers}
\title{Internal helper for plotting workflows}
\usage{
.plot_network_build_ggraph_layers(
  lay,
  palette_vals,
  node_size,
  edge_width,
  label_size,
  max.overlaps,
  hub_border_col,
  hub_border_size,
  title = NULL,
  caption = NULL,
  linetype_name = NULL,
  linetype_values = NULL,
  linetype_breaks = NULL,
  linetype_labels = NULL,
  linetype_guide = NULL
)
}
\arguments{
\item{lay}{Parameter value.}

\item{palette_vals}{Parameter value.}

\item{node_size}{Parameter value.}

\item{edge_width}{Parameter value.}

\item{label_size}{Parameter value.}

\item{max.overlaps}{Parameter value.}

\item{hub_border_col}{Parameter value.}

\item{hub_border_size}{Parameter value.}

\item{title}{Parameter value.}

\item{caption}{Parameter value.}

\item{linetype_name}{Parameter value.}

\item{linetype_values}{Parameter value.}

\item{linetype_breaks}{Parameter value.}

\item{linetype_labels}{Parameter value.}

\item{linetype_guide}{Parameter value.}
}
\value{
Internal helper result
}
\description{
Internal helper for `.plot_network_build_ggraph_layers`.
}
\keyword{internal}
