% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.plot_network_prepare}
\alias{.plot_network_prepare}
\title{Internal helper for plotting workflows}
\usage{
.plot_network_prepare(...)
}
\arguments{
\item{...}{Additional arguments (currently unused).}
}
\value{
List carrying the plotting arguments
}
\description{
Internal helper for `.plot_network_prepare`.
}
\keyword{internal}
