% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.precision_at_k}
\alias{.precision_at_k}
\title{Precision at K for a score ranking.}
\usage{
.precision_at_k(labels, scores, k = 100L)
}
\description{
Precision at K for a score ranking.
}
\keyword{internal}
