% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.prefetch_roi_molecules}
\alias{.prefetch_roi_molecules}
\title{Prefetch Roi Molecules}
\usage{
.prefetch_roi_molecules(state)
}
\arguments{
\item{state}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.prefetch_roi_molecules`.
}
\keyword{internal}
