% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.py_module_available}
\alias{.py_module_available}
\title{Check if Python module is available (adapter)}
\usage{
.py_module_available(module)
}
\arguments{
\item{module}{Module name}
}
\value{
Logical
}
\description{
Wrapper around reticulate::py_module_available
}
\keyword{internal}
