% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.py_to_r}
\alias{.py_to_r}
\title{Convert Python object to R (adapter)}
\usage{
.py_to_r(py_obj)
}
\arguments{
\item{py_obj}{Python object}
}
\value{
R object
}
\description{
Wrapper around reticulate::py_to_r
}
\keyword{internal}
