% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.r_to_py}
\alias{.r_to_py}
\title{Convert R object to Python (adapter)}
\usage{
.r_to_py(r_obj, convert = TRUE)
}
\arguments{
\item{r_obj}{R object}

\item{convert}{Whether to convert on access}
}
\value{
Python object
}
\description{
Wrapper around reticulate::r_to_py
}
\keyword{internal}
