% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.read_cosmx_header}
\alias{.read_cosmx_header}
\title{Read Cosmx Header}
\usage{
.read_cosmx_header(expr_mat)
}
\arguments{
\item{expr_mat}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.read_cosmx_header`.
}
\keyword{internal}
