% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.read_hdf5}
\alias{.read_hdf5}
\title{Read HDF5 file (adapter)}
\usage{
.read_hdf5(file, name)
}
\arguments{
\item{file}{Path to HDF5 file}

\item{name}{Dataset name within file}
}
\value{
Matrix or array
}
\description{
Read data from HDF5 file using available package
}
\keyword{internal}
