% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_imaging.R
\name{.read_jpeg}
\alias{.read_jpeg}
\title{Read JPEG image (adapter)}
\usage{
.read_jpeg(source, native = FALSE)
}
\arguments{
\item{source}{Path to JPEG file}

\item{native}{Whether to return native raster (default FALSE)}
}
\value{
Raster array
}
\description{
Wrapper around jpeg::readJPEG
}
\keyword{internal}
