% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_imaging.R
\name{.read_png}
\alias{.read_png}
\title{Read PNG image (adapter)}
\usage{
.read_png(source, native = FALSE, info = FALSE)
}
\arguments{
\item{source}{Path to PNG file}

\item{native}{Whether to return native raster (default FALSE)}

\item{info}{Whether to return image info (default FALSE)}
}
\value{
Raster array
}
\description{
Wrapper around png::readPNG
}
\keyword{internal}
