% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.read_scalefactors_json}
\alias{.read_scalefactors_json}
\title{Read Scalefactors Json}
\usage{
.read_scalefactors_json(json_path, warn_if_missing = TRUE)
}
\arguments{
\item{json_path}{Filesystem path.}

\item{warn_if_missing}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.read_scalefactors_json`.
}
\keyword{internal}
