% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.read_xenium_sparse_components}
\alias{.read_xenium_sparse_components}
\title{Read Xenium Sparse Components}
\usage{
.read_xenium_sparse_components(h5_file)
}
\arguments{
\item{h5_file}{Filesystem path.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.read_xenium_sparse_components`.
}
\keyword{internal}
