% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.reorder_hclust}
\alias{.reorder_hclust}
\title{Reorder hclust (adapter)}
\usage{
.reorder_hclust(hclust_obj, dist)
}
\arguments{
\item{hclust_obj}{hclust object}

\item{dist}{Distance matrix}
}
\value{
Reordered hclust object
}
\description{
Reorder hierarchical clustering using gclus
}
\keyword{internal}
