% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.require_bigmemory_or_stop}
\alias{.require_bigmemory_or_stop}
\title{Require bigmemory or stop}
\usage{
.require_bigmemory_or_stop()
}
\value{
Invisible TRUE
}
\description{
Require bigmemory or stop
}
\keyword{internal}
