% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_parallel.R
\name{.require_future_or_stop}
\alias{.require_future_or_stop}
\title{Require future or stop}
\usage{
.require_future_or_stop()
}
\value{
Invisible TRUE
}
\description{
Require future or stop
}
\keyword{internal}
