% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.require_hdf5_or_stop}
\alias{.require_hdf5_or_stop}
\title{Require HDF5 support or stop}
\usage{
.require_hdf5_or_stop()
}
\value{
Invisible TRUE
}
\description{
Require HDF5 support or stop
}
\keyword{internal}
