% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_imaging.R
\name{.require_jpeg_or_stop}
\alias{.require_jpeg_or_stop}
\title{Require jpeg or stop}
\usage{
.require_jpeg_or_stop()
}
\value{
Invisible TRUE
}
\description{
Require jpeg or stop
}
\keyword{internal}
