% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.require_pyarrow_or_stop}
\alias{.require_pyarrow_or_stop}
\title{Require pyarrow or stop}
\usage{
.require_pyarrow_or_stop()
}
\value{
Invisible TRUE
}
\description{
Require pyarrow or stop
}
\keyword{internal}
