% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.require_reticulate_or_stop}
\alias{.require_reticulate_or_stop}
\title{Require reticulate or stop}
\usage{
.require_reticulate_or_stop()
}
\value{
Invisible TRUE
}
\description{
Require reticulate or stop
}
\keyword{internal}
