% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.resolve_histology_y_origin}
\alias{.resolve_histology_y_origin}
\title{Resolve Histology Y Origin}
\usage{
.resolve_histology_y_origin(grid_layer, y_origin_choice)
}
\arguments{
\item{grid_layer}{Layer name.}

\item{y_origin_choice}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.resolve_histology_y_origin`.
}
\keyword{internal}
