% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.resolve_xenium_matrix_path}
\alias{.resolve_xenium_matrix_path}
\title{Resolve Xenium Matrix Path}
\usage{
.resolve_xenium_matrix_path(xenium_dir)
}
\arguments{
\item{xenium_dir}{Filesystem path.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.resolve_xenium_matrix_path`.
}
\keyword{internal}
