% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.reticulate_available}
\alias{.reticulate_available}
\title{Check if reticulate is available}
\usage{
.reticulate_available()
}
\value{
Logical
}
\description{
Check if reticulate is available
}
\keyword{internal}
