% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.rotate_dendrogram}
\alias{.rotate_dendrogram}
\title{Rotate dendrogram (adapter)}
\usage{
.rotate_dendrogram(dend, ...)
}
\arguments{
\item{dend}{Dendrogram object}

\item{...}{Additional arguments}
}
\value{
Rotated dendrogram
}
\description{
Rotate dendrogram using dendextend
}
\keyword{internal}
