% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_graph_core.R
\name{.run_graph_algorithm}
\alias{.run_graph_algorithm}
\title{Run Graph Algorithm}
\usage{
.run_graph_algorithm(g, algo_name, res_param, objective)
}
\arguments{
\item{g}{Parameter value.}

\item{algo_name}{Parameter value.}

\item{res_param}{Parameter value.}

\item{objective}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.run_graph_algorithm`.
}
\keyword{internal}
