% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.run_single_partition}
\alias{.run_single_partition}
\title{Run Single Partition}
\usage{
.run_single_partition(
  graph,
  backend = c("igraph", "networkit"),
  algo_name = c("leiden", "louvain"),
  res_param,
  objective,
  threads = 1L,
  nk_opts = list()
)
}
\arguments{
\item{graph}{Parameter value.}

\item{backend}{Parameter value.}

\item{algo_name}{Parameter value.}

\item{res_param}{Parameter value.}

\item{objective}{Parameter value.}

\item{threads}{Number of threads to use.}

\item{nk_opts}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.run_single_partition`.
}
\keyword{internal}
