% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_modules_stage4.R
\name{.run_stats_table_postprocess_module}
\alias{.run_stats_table_postprocess_module}
\title{Run Stats Table Postprocess Module}
\usage{
.run_stats_table_postprocess_module(fixture)
}
\arguments{
\item{fixture}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.run_stats_table_postprocess_module`.
}
\keyword{internal}
