% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.run_xenium_pipeline}
\alias{.run_xenium_pipeline}
\title{Run Xenium Pipeline}
\usage{
.run_xenium_pipeline(cfg)
}
\arguments{
\item{cfg}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.run_xenium_pipeline`.
}
\keyword{internal}
