% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.safe_set_dim}
\alias{.safe_set_dim}
\title{Safe Set Dim}
\usage{
.safe_set_dim(M, target_dim)
}
\arguments{
\item{M}{Parameter value.}

\item{target_dim}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.safe_set_dim`.
}
\keyword{internal}
