% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.safe_set_dimnames}
\alias{.safe_set_dimnames}
\title{Safe Set Dimnames}
\usage{
.safe_set_dimnames(M, target_dn)
}
\arguments{
\item{M}{Parameter value.}

\item{target_dn}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.safe_set_dimnames`.
}
\keyword{internal}
