% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.safe_thread_count}
\alias{.safe_thread_count}
\title{Safe Thread Count}
\usage{
.safe_thread_count()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.safe_thread_count`.
}
\keyword{internal}
