% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_stats_correlation.R
\name{.select_top_genes}
\alias{.select_top_genes}
\title{Intelligent gene subset selection}
\usage{
.select_top_genes(mat, max_genes = 50000, method = "variance", verbose = TRUE)
}
\arguments{
\item{mat}{Expression matrix}

\item{max_genes}{Maximum number of genes}

\item{method}{Selection method: "variance", "mean", "dropout", "random"}

\item{verbose}{Logical.}
}
\value{
Selected gene names vector
}
\description{
Select gene subset intelligently based on dataset size and available memory
}
\keyword{internal}
