% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.Pearson.r
\name{.selectTopGenes}
\alias{.selectTopGenes}
\title{Intelligent gene subset selection}
\usage{
.selectTopGenes(mat, max_genes = 50000, method = "variance")
}
\arguments{
\item{mat}{Expression matrix}

\item{max_genes}{Maximum number of genes}

\item{method}{Selection method: "variance", "mean", "dropout", "random"}
}
\value{
Selected gene names vector
}
\description{
Select gene subset intelligently based on dataset size and available memory
}
