% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.seriate_matrix}
\alias{.seriate_matrix}
\title{Seriate matrix (adapter)}
\usage{
.seriate_matrix(mat, method = "PCA")
}
\arguments{
\item{mat}{Matrix to seriate}

\item{method}{Seriation method (default "PCA")}
}
\value{
Seriated matrix
}
\description{
Reorder matrix using seriation
}
\keyword{internal}
