% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.seriation_available}
\alias{.seriation_available}
\title{Check if seriation is available}
\usage{
.seriation_available()
}
\value{
Logical
}
\description{
Check if seriation is available
}
\keyword{internal}
