% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.set_blas_environment}
\alias{.set_blas_environment}
\title{Set Blas Environment}
\usage{
.set_blas_environment()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.set_blas_environment`.
}
\keyword{internal}
