% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.set_conservative_threads}
\alias{.set_conservative_threads}
\title{Set Conservative Threads}
\usage{
.set_conservative_threads()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.set_conservative_threads`.
}
\keyword{internal}
