% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_visium.R
\name{.set_seurat_default_assay}
\alias{.set_seurat_default_assay}
\title{Set default assay in Seurat object (adapter)}
\usage{
.set_seurat_default_assay(seurat_obj, assay)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{assay}{Assay name}
}
\value{
Seurat object with updated default assay
}
\description{
Wrapper around Seurat::DefaultAssay
}
\keyword{internal}
