% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{.sort_dendrogram}
\alias{.sort_dendrogram}
\title{Sort dendrogram (adapter)}
\usage{
.sort_dendrogram(dend)
}
\arguments{
\item{dend}{Dendrogram object}
}
\value{
Sorted dendrogram
}
\description{
Sort dendrogram using dendsort
}
\keyword{internal}
