% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_stats_correlation.R
\name{.sparse_to_dense_chunked}
\alias{.sparse_to_dense_chunked}
\title{Chunked conversion of sparse matrix to dense matrix}
\usage{
.sparse_to_dense_chunked(sparse_mat, chunk_size)
}
\arguments{
\item{sparse_mat}{Sparse matrix}

\item{chunk_size}{Chunk size}
}
\value{
Dense matrix
}
\description{
Internal helper for `.sparse_to_dense_chunked`.
}
\keyword{internal}
