% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.Pearson.r
\name{.sparseToDenseChunked}
\alias{.sparseToDenseChunked}
\title{Chunked conversion of sparse matrix to dense matrix}
\usage{
.sparseToDenseChunked(sparse_mat, chunk_size)
}
\arguments{
\item{sparse_mat}{Sparse matrix}

\item{chunk_size}{Chunk size}
}
\value{
Dense matrix
}
\description{
Chunked conversion of sparse matrix to dense matrix
}
