% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.spatial_within}
\alias{.spatial_within}
\title{Spatial predicate: within (adapter)}
\usage{
.spatial_within(x, y)
}
\arguments{
\item{x}{sf object}

\item{y}{sf object}
}
\value{
Logical matrix
}
\description{
Test if geometries are within other geometries
}
\keyword{internal}
