% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_spatial_weights.R
\name{.spw_analyse_hex_geometry}
\alias{.spw_analyse_hex_geometry}
\title{Spw Analyse Hex Geometry}
\usage{
.spw_analyse_hex_geometry(gi)
}
\arguments{
\item{gi}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.spw_analyse_hex_geometry`.
}
\keyword{internal}
