% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_spatial_weights.R
\name{.spw_hex_col_parity}
\alias{.spw_hex_col_parity}
\title{Spw Hex Col Parity}
\usage{
.spw_hex_col_parity(gi)
}
\arguments{
\item{gi}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.spw_hex_col_parity`.
}
\keyword{internal}
