% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_spatial_weights.R
\name{.spw_infer_hex_variant}
\alias{.spw_infer_hex_variant}
\title{Spw Infer Hex Variant}
\usage{
.spw_infer_hex_variant(grid_info, force = FALSE)
}
\arguments{
\item{grid_info}{Parameter value.}

\item{force}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.spw_infer_hex_variant`.
}
\keyword{internal}
