% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_spatial_weights.R
\name{.spw_is_integerish}
\alias{.spw_is_integerish}
\title{Spw Is Integerish}
\usage{
.spw_is_integerish(value, tol = 1e-06)
}
\arguments{
\item{value}{Parameter value.}

\item{tol}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.spw_is_integerish`.
}
\keyword{internal}
