% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_spatial_weights.R
\name{.spw_should_enable_parallel_nb}
\alias{.spw_should_enable_parallel_nb}
\title{Spw Should Enable Parallel Nb}
\usage{
.spw_should_enable_parallel_nb(
  ncores,
  verbose,
  parent = "computeWeights",
  step = "S03"
)
}
\arguments{
\item{ncores}{Number of cores/threads to use.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.spw_should_enable_parallel_nb`.
}
\keyword{internal}
