% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_spatial_weights.R
\name{.spw_trace_enabled}
\alias{.spw_trace_enabled}
\title{Spw Trace Enabled}
\usage{
.spw_trace_enabled()
}
\value{
Return value used internally.
}
\description{
Internal helper for `.spw_trace_enabled`.
}
\keyword{internal}
