% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.stage1_postprocess_and_assemble}
\alias{.stage1_postprocess_and_assemble}
\title{Stage1 Postprocess And Assemble}
\usage{
.stage1_postprocess_and_assemble(
  kept_genes,
  genes_all,
  threshold,
  similarity_sub,
  similarity_graph,
  mode_selected,
  restart_memberships,
  stage1_consensus_matrix,
  stage1_membership_labels,
  stage1_backend,
  stage1_algo,
  stage1_algo_per_run
)
}
\arguments{
\item{kept_genes}{Parameter value.}

\item{genes_all}{Parameter value.}

\item{threshold}{Parameter value.}

\item{similarity_sub}{Parameter value.}

\item{similarity_graph}{Parameter value.}

\item{mode_selected}{Parameter value.}

\item{restart_memberships}{Parameter value.}

\item{stage1_consensus_matrix}{Parameter value.}

\item{stage1_membership_labels}{Parameter value.}

\item{stage1_backend}{Parameter value.}

\item{stage1_algo}{Parameter value.}

\item{stage1_algo_per_run}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.stage1_postprocess_and_assemble`.
}
\keyword{internal}
