% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.stage2_configure_runtime_threads}
\alias{.stage2_configure_runtime_threads}
\title{Stage2 Configure Runtime Threads}
\usage{
.stage2_configure_runtime_threads(config, stage1_mode_selected, kept_genes)
}
\arguments{
\item{config}{Parameter value.}

\item{stage1_mode_selected}{Parameter value.}

\item{kept_genes}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.stage2_configure_runtime_threads`.
}
\keyword{internal}
