% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_modules_stage4.R
\name{.stage2_native_inputs_spec_build}
\alias{.stage2_native_inputs_spec_build}
\title{Stage2 Native Inputs Spec Build}
\usage{
.stage2_native_inputs_spec_build(
  validated_inputs,
  config,
  stage1_membership_labels
)
}
\arguments{
\item{validated_inputs}{Parameter value.}

\item{config}{Parameter value.}

\item{stage1_membership_labels}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.stage2_native_inputs_spec_build`.
}
\keyword{internal}
