% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.stage2_postprocess_and_assemble}
\alias{.stage2_postprocess_and_assemble}
\title{Stage2 Postprocess And Assemble}
\usage{
.stage2_postprocess_and_assemble(
  memb_final,
  cons,
  metrics_before,
  genes_before,
  edges_corr,
  corrected_similarity_graph,
  W,
  kept_genes,
  genes_all,
  stage1_membership_matrix,
  stage1_consensus_matrix,
  stage1_similarity_graph,
  similarity_matrix,
  use_log1p_weight,
  runtime_cfg,
  config,
  L_post,
  stage2_backend
)
}
\arguments{
\item{memb_final}{Parameter value.}

\item{cons}{Parameter value.}

\item{metrics_before}{Parameter value.}

\item{genes_before}{Parameter value.}

\item{edges_corr}{Parameter value.}

\item{corrected_similarity_graph}{Parameter value.}

\item{W}{Parameter value.}

\item{kept_genes}{Parameter value.}

\item{genes_all}{Parameter value.}

\item{stage1_membership_matrix}{Parameter value.}

\item{stage1_consensus_matrix}{Parameter value.}

\item{stage1_similarity_graph}{Parameter value.}

\item{similarity_matrix}{Parameter value.}

\item{use_log1p_weight}{Logical flag.}

\item{runtime_cfg}{Parameter value.}

\item{config}{Parameter value.}

\item{L_post}{Parameter value.}

\item{stage2_backend}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.stage2_postprocess_and_assemble`.
}
\keyword{internal}
