% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.standardize_crop_bbox}
\alias{.standardize_crop_bbox}
\title{Standardize Crop Bbox}
\usage{
.standardize_crop_bbox(bbox, warn_if_missing_names = TRUE)
}
\arguments{
\item{bbox}{Parameter value.}

\item{warn_if_missing_names}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.standardize_crop_bbox`.
}
\keyword{internal}
