% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_modules_stage4.R
\name{.stats_table_postprocess_spec_build}
\alias{.stats_table_postprocess_spec_build}
\title{Stats Table Postprocess Spec Build}
\usage{
.stats_table_postprocess_spec_build(validated_inputs)
}
\arguments{
\item{validated_inputs}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.stats_table_postprocess_spec_build`.
}
\keyword{internal}
