% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_external_reference.R
\name{.stringdb_available}
\alias{.stringdb_available}
\title{Check if STRINGdb is available}
\usage{
.stringdb_available()
}
\value{
Logical indicating if STRINGdb is installed
}
\description{
Check if STRINGdb is available
}
\keyword{internal}
