% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.stringdb_cache_env}
\alias{.stringdb_cache_env}
\title{Access package-level cache environment for STRINGdb.}
\usage{
.stringdb_cache_env()
}
\description{
Access package-level cache environment for STRINGdb.
}
\keyword{internal}
