% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.stringdb_cache_get}
\alias{.stringdb_cache_get}
\title{Retrieve cached STRINGdb artifacts.}
\usage{
.stringdb_cache_get(scope_obj = NULL, cache_dir = NULL, cache_type, cache_key)
}
\description{
Retrieve cached STRINGdb artifacts.
}
\keyword{internal}
