% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.stringdb_cache_put}
\alias{.stringdb_cache_put}
\title{Store cached STRINGdb artifacts.}
\usage{
.stringdb_cache_put(
  scope_obj = NULL,
  cache_dir = NULL,
  cache_type,
  cache_key,
  value
)
}
\description{
Store cached STRINGdb artifacts.
}
\keyword{internal}
