% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_connect}
\alias{.stringdb_connect}
\title{Create STRINGdb instance.}
\usage{
.stringdb_connect(
  species = 9606,
  score_threshold = 0,
  version = "11.5",
  cache_dir = NULL
)
}
\description{
Create STRINGdb instance.
}
\keyword{internal}
