% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_get_interactions}
\alias{.stringdb_get_interactions}
\title{Fetch STRING interactions for provided STRING IDs with optional caching.}
\usage{
.stringdb_get_interactions(string_db, string_ids, cache_dir = NULL)
}
\description{
Fetch STRING interactions for provided STRING IDs with optional caching.
}
\keyword{internal}
