% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.stringdb_get_interactions_cached}
\alias{.stringdb_get_interactions_cached}
\title{Fetch STRING interactions with caching.}
\usage{
.stringdb_get_interactions_cached(
  string_db,
  string_ids,
  score_threshold,
  scope_obj = NULL,
  cache_dir = NULL
)
}
\description{
Fetch STRING interactions with caching.
}
\keyword{internal}
