% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.stringdb_interaction_paths}
\alias{.stringdb_interaction_paths}
\title{Build STRINGdb interaction URL and local file path.}
\usage{
.stringdb_interaction_paths(string_db)
}
\description{
Build STRINGdb interaction URL and local file path.
}
\keyword{internal}
