% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_label_edges}
\alias{.stringdb_label_edges}
\title{Annotate edges with STRING scores and labels.}
\usage{
.stringdb_label_edges(
  edge_df,
  gene_to_string,
  interactions,
  score_threshold = 700
)
}
\description{
Annotate edges with STRING scores and labels.
}
\keyword{internal}
