% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_map_genes}
\alias{.stringdb_map_genes}
\title{Map genes to STRING IDs with optional caching.}
\usage{
.stringdb_map_genes(string_db, genes, cache_dir = NULL)
}
\description{
Map genes to STRING IDs with optional caching.
}
\keyword{internal}
