% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_store.R
\name{.stringdb_map_genes_cached}
\alias{.stringdb_map_genes_cached}
\title{Map genes to STRING IDs with caching.}
\usage{
.stringdb_map_genes_cached(
  string_db,
  genes,
  input_id_type = "gene",
  scope_obj = NULL,
  cache_dir = NULL
)
}
\description{
Map genes to STRING IDs with caching.
}
\keyword{internal}
