% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_prepare_cache_files}
\alias{.stringdb_prepare_cache_files}
\title{Prepare STRINGdb download cache with md5 checks.}
\usage{
.stringdb_prepare_cache_files(string_db, cache_dir, force_update = NULL)
}
\description{
Prepare STRINGdb download cache with md5 checks.
}
\keyword{internal}
