% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_require_or_stop}
\alias{.stringdb_require_or_stop}
\title{Validate STRINGdb availability.}
\usage{
.stringdb_require_or_stop()
}
\description{
Validate STRINGdb availability.
}
\keyword{internal}
