% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_external_reference_stringdb.R
\name{.stringdb_update_flag}
\alias{.stringdb_update_flag}
\title{Read update flag from env/option.}
\usage{
.stringdb_update_flag()
}
\description{
Read update flag from env/option.
}
\keyword{internal}
