% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.symmetric_pmax}
\alias{.symmetric_pmax}
\title{Symmetric Pmax}
\usage{
.symmetric_pmax(M)
}
\arguments{
\item{M}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.symmetric_pmax`.
}
\keyword{internal}
