% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_spatial_geometry.R
\name{.union_geometries}
\alias{.union_geometries}
\title{Union geometries (adapter)}
\usage{
.union_geometries(x, y = NULL)
}
\arguments{
\item{x}{sf object}

\item{y}{Optional second sf object}
}
\value{
Unioned geometry
}
\description{
Wrapper around sf::st_union
}
\keyword{internal}
