% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.use_conda_environment}
\alias{.use_conda_environment}
\title{Use Conda Environment}
\usage{
.use_conda_environment(env_name, conda_bin = NULL, required = FALSE)
}
\arguments{
\item{env_name}{Parameter value.}

\item{conda_bin}{Parameter value.}

\item{required}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.use_conda_environment`.
}
\keyword{internal}
