% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_python_bridge.R
\name{.use_python_env}
\alias{.use_python_env}
\title{Use specific Python environment (adapter)}
\usage{
.use_python_env(env_type = c("conda", "virtualenv", "python"), env_name)
}
\arguments{
\item{env_type}{Environment type ("conda", "virtualenv", "python")}

\item{env_name}{Environment name or path}
}
\value{
Invisible NULL
}
\description{
Set Python environment for reticulate
}
\keyword{internal}
