% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.use_python_interpreter}
\alias{.use_python_interpreter}
\title{Use Python Interpreter}
\usage{
.use_python_interpreter(python_path, required = FALSE)
}
\arguments{
\item{python_path}{Filesystem path.}

\item{required}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.use_python_interpreter`.
}
\keyword{internal}
