% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pipeline.R
\name{.use_virtual_environment}
\alias{.use_virtual_environment}
\title{Use Virtual Environment}
\usage{
.use_virtual_environment(env_name, required = FALSE)
}
\arguments{
\item{env_name}{Parameter value.}

\item{required}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.use_virtual_environment`.
}
\keyword{internal}
