% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.visium_build_grid_components}
\alias{.visium_build_grid_components}
\title{Visium Build Grid Components}
\usage{
.visium_build_grid_components(
  pos_dt,
  grid_multiplier,
  microns_per_pixel,
  raw_counts,
  sct_mat,
  data_type
)
}
\arguments{
\item{pos_dt}{Parameter value.}

\item{grid_multiplier}{Parameter value.}

\item{microns_per_pixel}{Parameter value.}

\item{raw_counts}{Parameter value.}

\item{sct_mat}{Parameter value.}

\item{data_type}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_build_grid_components`.
}
\keyword{internal}
