% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_visium.R
\name{.visium_deps_available}
\alias{.visium_deps_available}
\title{Check if Visium dependencies are available}
\usage{
.visium_deps_available()
}
\value{
Logical indicating if all Visium dependencies are installed
}
\description{
Check if Visium dependencies are available
}
\keyword{internal}
