% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.visium_load_positions}
\alias{.visium_load_positions}
\title{Visium Load Positions}
\usage{
.visium_load_positions(
  spatial_dir,
  include_in_tissue_spots,
  raw_counts,
  microns_per_pixel,
  flip_y
)
}
\arguments{
\item{spatial_dir}{Filesystem path.}

\item{include_in_tissue_spots}{Parameter value.}

\item{raw_counts}{Parameter value.}

\item{microns_per_pixel}{Parameter value.}

\item{flip_y}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_load_positions`.
}
\keyword{internal}
