% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.visium_load_scalefactors}
\alias{.visium_load_scalefactors}
\title{Visium Load Scalefactors}
\usage{
.visium_load_scalefactors(spatial_dir, verbose)
}
\arguments{
\item{spatial_dir}{Filesystem path.}

\item{verbose}{Logical; whether to emit progress messages.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_load_scalefactors`.
}
\keyword{internal}
