% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.visium_load_seurat_data}
\alias{.visium_load_seurat_data}
\title{Visium Load Seurat Data}
\usage{
.visium_load_seurat_data(
  input_dir,
  run_sctransform,
  sct_return_only_var_genes,
  vars_to_regress,
  glmGamPoi,
  verbose,
  data_type
)
}
\arguments{
\item{input_dir}{Filesystem path.}

\item{run_sctransform}{Parameter value.}

\item{sct_return_only_var_genes}{Parameter value.}

\item{vars_to_regress}{Parameter value.}

\item{glmGamPoi}{Parameter value.}

\item{verbose}{Logical; whether to emit progress messages.}

\item{data_type}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_load_seurat_data`.
}
\keyword{internal}
