% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_modules_stage4.R
\name{.visium_scope_spec_build}
\alias{.visium_scope_spec_build}
\title{Visium Scope Spec Build}
\usage{
.visium_scope_spec_build(validated_inputs)
}
\arguments{
\item{validated_inputs}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_scope_spec_build`.
}
\keyword{internal}
