% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_modules_stage4.R
\name{.visium_scope_validate_inputs}
\alias{.visium_scope_validate_inputs}
\title{Visium Scope Validate Inputs}
\usage{
.visium_scope_validate_inputs(
  input_dir,
  grid_multiplier,
  include_in_tissue_spots,
  flip_y,
  data_type
)
}
\arguments{
\item{input_dir}{Filesystem path.}

\item{grid_multiplier}{Parameter value.}

\item{include_in_tissue_spots}{Parameter value.}

\item{flip_y}{Parameter value.}

\item{data_type}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_scope_validate_inputs`.
}
\keyword{internal}
