% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ingest.R
\name{.visium_validate_inputs}
\alias{.visium_validate_inputs}
\title{Visium Validate Inputs}
\usage{
.visium_validate_inputs(input_dir, grid_multiplier, data_type)
}
\arguments{
\item{input_dir}{Filesystem path.}

\item{grid_multiplier}{Parameter value.}

\item{data_type}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.visium_validate_inputs`.
}
\keyword{internal}
