% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_big_data.R
\name{.write_hdf5}
\alias{.write_hdf5}
\title{Write HDF5 file (adapter)}
\usage{
.write_hdf5(data, file, name)
}
\arguments{
\item{data}{Data to write}

\item{file}{Path to HDF5 file}

\item{name}{Dataset name within file}
}
\value{
Invisible NULL
}
\description{
Write data to HDF5 file using available package
}
\keyword{internal}
