% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{fixRNG}
\alias{fixRNG}
\title{Fix Random Number Generation}
\usage{
fixRNG(seed = 42)
}
\arguments{
\item{seed}{Integer. Random seed to set}
}
\description{
Ensures reproducible random number generation across platforms
}
\keyword{internal}
