% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearson.r
\name{geneCorrelation}
\alias{geneCorrelation}
\title{Compute gene × gene correlation and save to CoordObj
Compute gene × gene correlation and (optionally) save to CoordObj}
\usage{
geneCorrelation(
  coordObj,
  grid_name = "grid_lenGrid16",
  layer = "Xz",
  genes = NULL,
  method = "pearson",
  blocksize = 2000,
  as_sparse = TRUE,
  thr = 0.3,
  use_parallel = FALSE,
  ncores = 15,
  cor_layer_name = NULL,
  write_back = TRUE
)
}
\arguments{
\item{cor_layer_name}{Character; name of the new sub-layer to save.
Default = paste0(method, "_cor")}

\item{write_back}{Logical; if TRUE, write matrix to
coordObj@grid[\link{grid_name}][\link{cor_layer_name}]}
}
\value{
If write_back = TRUE, the modified CoordObj; otherwise the
correlation matrix (dense or dgCMatrix)
}
\description{
Compute gene × gene correlation and save to CoordObj
Compute gene × gene correlation and (optionally) save to CoordObj
}
