% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG2CLI-package.R, R/zzz.r
\docType{package}
\name{geneSCOPE-package}
\alias{geneSCOPE}
\alias{geneSCOPE-package}
\title{geneSCOPE}
\description{
Spatial gene correlation analysis toolkit. This file provides package-level metadata and
useDynLib registration hints.

Package initialization functions and utility helpers for cross-platform
compatibility and safe thread management.
}
\details{
Package Loading and Environment Setup
}
\author{
\strong{Maintainer}: Shicheng Zhang \email{0322704@ed.tus.ac.jp}

Authors:
\itemize{
  \item Koichi Saeki
  \item Hiroshi Haeno
}

}
\keyword{internal}
