% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_threading.r, R/zzz.r
\name{getSafeThreadCount}
\alias{getSafeThreadCount}
\title{Get safe thread count}
\usage{
getSafeThreadCount(requested_cores)

getSafeThreadCount(requested_cores)
}
\arguments{
\item{requested_cores}{Integer. Requested number of cores}

\item{max_requested}{Maximum threads requested}
}
\value{
Safe number of threads to use

Integer. Safe number of cores to use
}
\description{
Returns a conservative thread count that should work on most systems

Determines safe thread count based on platform and available cores
}
\keyword{internal}
