% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.PlotNetWork.r
\name{getValidClusterColumns}
\alias{getValidClusterColumns}
\title{获取有效的聚类列名}
\usage{
getValidClusterColumns(
  scope_obj,
  grid_size = 10,
  include_cmh = FALSE,
  pct_filters = c("All", "q85.0", "q90.0", "q95.0", "q99.0", "q99.9")
)
}
\arguments{
\item{scope_obj}{scope_object对象}

\item{grid_size}{网格尺寸}

\item{include_cmh}{是否包含cmh相关列}

\item{pct_filters}{百分比过滤字符串向量}
}
\value{
存在于scope_obj中的有效列名向量
}
\description{
根据网格尺寸和聚类参数模式生成有效的列名列表
}
