% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_spatial_weights.R
\name{grid_weights_kernel_hexr_omp}
\alias{grid_weights_kernel_hexr_omp}
\title{grid_weights_kernel_hexr_omp - R fallback}
\usage{
grid_weights_kernel_hexr_omp(
  nrow,
  ncol,
  gx,
  gy,
  oddr = TRUE,
  radius = 2L,
  kernel = "gaussian",
  sigma = 1
)
}
\description{
grid_weights_kernel_hexr_omp - R fallback
}
\keyword{internal}
