% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.r
\name{idelta}
\alias{idelta}
\title{Compute Morisita's Index Delta}
\usage{
idelta(G, n_threads = 1)
}
\arguments{
\item{G}{Sparse count matrix (genes x grids)}

\item{n_threads}{Number of threads}
}
\value{
Vector of Morisita's delta values
}
\description{
Computes Morisita's index of dispersion for sparse count matrix.
}
