% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.r
\name{nb2mat}
\alias{nb2mat}
\title{Convert Neighbor List to Sparse Matrix}
\usage{
nb2mat(nb)
}
\arguments{
\item{nb}{Neighborhood list object}
}
\value{
A dgCMatrix sparse matrix
}
\description{
Transforms a neighborhood list into a binary sparse matrix (dgCMatrix).
}
