% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_advanced_viz.R
\name{optional-advanced-viz}
\alias{optional-advanced-viz}
\title{Advanced Visualization Optional Module}
\description{
Adapter layer for advanced visualization features.
  Requires ComplexHeatmap, circlize, ggfx, seriation, dendextend, dendsort, gclus.
}
\keyword{internal}
