% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.r
\name{pearson_cor}
\alias{pearson_cor}
\title{Compute Pearson Correlation}
\usage{
pearson_cor(X, bs = 2000, n_threads = 1)
}
\arguments{
\item{X}{Numeric matrix (already centered)}

\item{bs}{Block size for processing}

\item{n_threads}{Number of threads}
}
\value{
Correlation matrix
}
\description{
Computes block-wise Pearson correlation matrix in parallel.
}
