% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.PlotNetWork.r
\name{robustClusterCompare}
\alias{robustClusterCompare}
\title{执行健壮的聚类比较}
\usage{
robustClusterCompare(
  coordObj,
  grid_size = 10,
  include_cmh = FALSE,
  unify_types = TRUE,
  pct_filters = c("All", "q85.0", "q90.0", "q95.0", "q99.0", "q99.9"),
  ...
)
}
\arguments{
\item{coordObj}{CoordObj对象}

\item{grid_size}{网格尺寸}

\item{include_cmh}{是否包含cmh相关列}

\item{unify_types}{是否先统一数据类型}

\item{pct_filters}{百分比过滤字符串向量}

\item{...}{传递给plotClusterComparison的其他参数}
}
\value{
plotClusterComparison的返回值
}
\description{
统一数据类型并执行聚类比较
}
