% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{selectGridLayer}
\alias{selectGridLayer}
\title{Select Grid Layer}
\usage{
selectGridLayer(scope_obj, grid_name = NULL)
}
\arguments{
\item{scope_obj}{A scope_object}

\item{grid_name}{Optional grid layer name}
}
\value{
The selected grid layer
}
\description{
Helper function to select a grid layer from scope_obj
}
\keyword{internal}
