% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.PlotNetWork.r
\name{unifyClusteringTypes}
\alias{unifyClusteringTypes}
\title{统一聚类结果数据类型}
\usage{
unifyClusteringTypes(scope_obj, grid_sizes = c(10, 30, 55), verbose = TRUE)
}
\arguments{
\item{scope_obj}{scope_object对象，包含聚类结果}

\item{grid_sizes}{要处理的网格尺寸向量}

\item{verbose}{是否打印详细信息}
}
\value{
修改后的scope_object对象
}
\description{
将scope_object中指定的聚类结果列统一转换为因子类型
}
