% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_classes.r
\name{upgradeCellsSlot}
\alias{upgradeCellsSlot}
\title{Upgrade \code{CoordObj} Cells Slot}
\usage{
upgradeCellsSlot(obj)
}
\arguments{
\item{obj}{A \code{CoordObj} object with the \code{cells} slot as a sparse matrix.}
}
\value{
The modified \code{CoordObj} with the \code{cells} slot upgraded to a list format.
}
\description{
Converts the \code{cells} slot from a sparse matrix to a list format.
}
