% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12.ClusterGenes.r
\name{clusterGenes}
\alias{clusterGenes}
\title{Gene Clustering (results written to coordObj@meta.data)}
\usage{
clusterGenes(
  coordObj,
  grid_name = NULL,
  lee_stats_layer = "LeeStats_Xz",
  L_min = 0,
  use_FDR = TRUE,
  FDR_max = 0.05,
  pct_min = "q0",
  drop_isolated = TRUE,
  algo = c("leiden", "louvain"),
  resolution = 1,
  objective = c("CPM", "modularity"),
  cluster_name = NULL,
  use_log1p_weight = TRUE,
  use_consensus = TRUE,
  graph_slot_name = "g_consensus",
  n_restart = 100,
  consensus_thr = 0,
  keep_cross_stable = TRUE,
  CI95_filter = FALSE,
  curve_layer = "LR_curve",
  CI_rule = c("remove_within", "remove_outside"),
  use_cmh_weight = FALSE,
  cmh_slot = "g_morisita",
  verbose = TRUE
)
}
\description{
Based on Lee's L statistics for filtering, graph construction, multiple community detection and optional consensus, output cluster numbers to coordObj@meta.data.
}
