% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearson.r
\name{computeDeltaLee}
\alias{computeDeltaLee}
\title{Compute Δ = Lee’s L − Pearson r with permutation p-values (+QC)}
\usage{
computeDeltaLee(
  coordObj,
  level = c("grid", "cell"),
  grid_name = "grid_lenGrid50",
  lee_stats_layer = "LeeStats_Xz",
  cor_layer_name = "pearson_Xz",
  user_delta = TRUE,
  r_threshold = 0.1
)
}
\arguments{
\item{level}{"grid" (default) or "cell".}

\item{grid_name}{Grid layer name (grid mode only).}

\item{lee_stats_layer}{Name of the Lee’s L sub-layer / attribute.}

\item{cor_layer_name}{Name of the Pearson correlation sub-layer / attribute.}

\item{perms}{Number of permutations.}

\item{fdr_method}{FDR adjustment method.}

\item{seed}{RNG seed.}
}
\value{
If successful, the modified \code{CoordObj}.
}
\description{
Calculates the difference Δ = L − r either on a grid layer
(\strong{level = "grid"}) or on the full single-cell matrix
(\strong{level = "cell"}), performs permutation testing via the
C++ backend \code{leeL_minusR_perm_ge_cpp()}, adds empirical
p-values, FDR, classification, and QC metrics, and writes the
result back into the \code{CoordObj}.
}
