% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_misc.R
\name{.clip_points_to_region}
\alias{.clip_points_to_region}
\title{Clip Points To Region}
\usage{
.clip_points_to_region(points, region, chunk_size = 5e+05, workers = 1L)
}
\arguments{
\item{points}{Parameter value.}

\item{region}{Parameter value.}

\item{chunk_size}{Parameter value.}

\item{workers}{Parameter value.}
}
\value{
Return value used internally.
}
\description{
Internal helper for `.clip_points_to_region`.
}
\keyword{internal}
