% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{selectGridLayer}
\alias{selectGridLayer}
\title{Select Grid Layer}
\usage{
selectGridLayer(coordObj, grid_name = NULL)
}
\arguments{
\item{coordObj}{A CoordObj}

\item{grid_name}{Optional grid layer name}
}
\value{
The selected grid layer
}
\description{
Helper function to select a grid layer from coordObj
}
\keyword{internal}
