% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_external_reference.R
\name{summarizeEvaluation}
\alias{summarizeEvaluation}
\title{Summarize edge and module evaluations.}
\usage{
summarizeEvaluation(edge_eval, module_eval)
}
\arguments{
\item{edge_eval}{Output from `evaluateExternalReference_STRING()`.}

\item{module_eval}{Output from `evaluateModuleQuality_STRING()`.}
}
\value{
A list with `edge_level_table` and `module_level_table`.
}
\description{
Summarize edge and module evaluations.
}
\examples{
\dontrun{
summary_tables <- summarizeEvaluation(edge_eval, mod_eval)
}
}
